### 2024-10-08 c767ccf
v0.7.2: Solve issue test_statistic equal to zero
