+++ phitter/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.7.1"
+__version__ = "0.7.2"
 
 from .main import PHITTER
 from phitter import continuous
