+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "phitter"
-version = "0.7.1"
+version = "0.7.2"
 description = "Find the best probability distribution for your dataset"
 authors = [{name = "Sebastián José Herrera Monterrosa", email = "phitter.email@gmail.com"}]
 readme = "README.md"
