+++ phitter/simulation/queueing_simulation/queueing_simulation.py
@@ -1,0 +1,939 @@
+import phitter
+import pandas as pd
+import random
+import numpy as np
+
+
+class QueueingSimulation:
+    def __init__(
+        self,
+        a: str,
+        a_parameters: dict,
+        s: str,
+        s_parameters: dict,
+        c: int,
+        k: float = float("inf"),
+        n: float = float("inf"),
+        d: str = "FIFO",
+        pbs_distribution: str | None = None,
+        pbs_parameters: dict | None = None,
+    ) -> None:
+        """Simulation of any queueing model.
+
+        Args:
+            a (str): Arrival time distribution (Arrival). Distributions that can be used: 'alpha', 'arcsine', 'argus', 'beta', 'beta_prime', 'beta_prime_4p', 'bradford', 'burr', 'burr_4p', 'cauchy', 'chi_square', 'chi_square_3p', 'dagum', 'dagum_4p', 'erlang', 'erlang_3p', 'error_function', 'exponential', 'exponential_2p', 'f', 'fatigue_life', 'folded_normal', 'frechet', 'f_4p', 'gamma', 'gamma_3p', 'generalized_extreme_value', 'generalized_gamma', 'generalized_gamma_4p', 'generalized_logistic', 'generalized_normal', 'generalized_pareto', 'gibrat', 'gumbel_left', 'gumbel_right', 'half_normal', 'hyperbolic_secant', 'inverse_gamma', 'inverse_gamma_3p', 'inverse_gaussian', 'inverse_gaussian_3p', 'johnson_sb', 'johnson_su', 'kumaraswamy', 'laplace', 'levy', 'loggamma', 'logistic', 'loglogistic', 'loglogistic_3p', 'lognormal', 'maxwell', 'moyal', 'nakagami', 'non_central_chi_square', 'non_central_f', 'non_central_t_student', 'normal', 'pareto_first_kind', 'pareto_second_kind', 'pert', 'power_function', 'rayleigh', 'reciprocal', 'rice', 'semicircular', 'trapezoidal', 'triangular', 't_student', 't_student_3p', 'uniform', 'weibull', 'weibull_3p', 'bernoulli', 'binomial', 'geometric', 'hypergeometric', 'logarithmic', 'negative_binomial', 'poisson'.
+            a_parameters (dict): All the needed parameters to use the probability distribution of the arrival time.
[…]
