### 2024-10-10 a26ea7e
Adding Additional Functions to QueueSim
