+++ phitter/simulation/queueing_simulation/queueing_simulation.py
@@ -4,6 +4,8 @@ import random
 import numpy as np
 import collections
 
+import math
+
 
 class QueueingSimulation:
     def __init__(
@@ -41,7 +43,7 @@ class QueueingSimulation:
             | phitter.discrete.DISCRETE_DISTRIBUTIONS
         )
 
-        self.__pbs_distribution = {
+        self.__pbs_distributions = {
             "own_distribution": OwnDistributions
         } | phitter.discrete.DISCRETE_DISTRIBUTIONS
 
@@ -49,15 +51,24 @@ class QueueingSimulation:
 
         self.__verify_variables(a, s, c, k, n, d, pbs_distribution, pbs_parameters)
 
-        self.__a = self.__probability_distribution[a](a_parameters)
-        self.__s = self.__probability_distribution[s](s_parameters)
+        self.__save_a = a
+        self.__save_a_params = a_parameters
+        self.__save_s = s
+        self.__save_s_params = s_parameters
+
+        self.__a = self.__probability_distribution[self.__save_a](self.__save_a_params)
+        self.__s = self.__probability_distribution[self.__save_s](self.__save_s_params)
         self.__c = c
         self.__k = k
         self.__n = n
         self.__d = d
 
+        self.__pbs_parameters = pbs_parameters
+        self.__pbs_distribution = pbs_distribution
         if d == "PBS":
-            self.__label = self.__pbs_distribution[pbs_distribution](pbs_parameters)
+            self.__label = self.__pbs_distributions[self.__pbs_distribution](
+                self.__pbs_parameters
+            )
 
         self.result_simulation = pd.DataFrame()
         self.__simulation_time = 0
@@ -130,9 +141,9 @@ class QueueingSimulation:
 
         if d == "PBS":
             if pbs_distribution != None and pbs_parameters != None:
-                if pbs_distribution not in self.__pbs_distribution:
+                if pbs_distribution not in self.__pbs_distributions:
                     raise ValueError(
-                        f"""You should select one of the following distributions: {self.__pbs_distribution}"""
+                        f"""You should select one of the following distributions: {self.__pbs_distributions}"""
                     )
             elif pbs_distribution == None and pbs_parameters == None:
                 raise ValueError(
@@ -143,9 +154,7 @@ class QueueingSimulation:
                 f"""You can only use 'pbs_distribution' and 'pbs_parameters' with 'd="PBS"'"""
             )
 
-    def run(
-        self, simulation_time: float = float("inf"), number_of_simulations: int = 1
-    ) -> tuple:
+    def run(self, simulation_time: float = float("inf")) -> tuple:
         """Simulation of any queueing model.
 
         Args:
@@ -161,11 +170,6 @@ class QueueingSimulation:
         if simulation_time <= 0:
             raise ValueError(
                 f"""'simulation_time' has to be a number and cannot be less or equals than zero."""
-            )
-
-        if number_of_simulations <= 0:
-            raise ValueError(
-                f"""'number_of_simulations' has to be a number and cannot be less or equals than zero."""
             )
 
         if simulation_time == float("inf"):
@@ -193,12 +197,17 @@ class QueueingSimulation:
         elif self.__d == "PBS":
             simulation_results = self.__pbs(simulation_time, simulation_results)
 
+        # Create a new column that is the same for all d
+        simulation_results["Finish after closed"] = [
+            1 if leave_time > self.__simulation_time else 0
+            for leave_time in simulation_results["Leave Time"]
+        ]
+
         self.__result_simulation = pd.DataFrame(simulation_results)
         self.__result_simulation = self.__result_simulation.drop(index=0).reset_index(
             drop=True
         )
 
-        # Probabilities
         self.elements_prob()
 
[…]
