+++ phitter/simulation/queueing_simulation/queueing_simulation.py
@@ -38,24 +38,28 @@ class QueueingSimulation:
             pbs_parameters (dict | None, optional): Parameters of the discrete distribution that identifies the label of the pbs, this parameter can only be used with "d='PBS'". If it is 'own-distribution' add labels in the following way (example): {0: 0.5, 1: 0.3, 2: 0.2}. Where the "key" corresponds to the label and the "value" the probability whose total sum must add up to 1; "keys" with greater importances are the smallers and always have to be numeric keys. You can add as labels as you need.
         """
 
+        # All phitter probability distributions
         self.__probability_distribution = (
             phitter.continuous.CONTINUOUS_DISTRIBUTIONS
             | phitter.discrete.DISCRETE_DISTRIBUTIONS
         )
 
+        # Distributions for labels
         self.__pbs_distributions = {
             "own_distribution": OwnDistributions
         } | phitter.discrete.DISCRETE_DISTRIBUTIONS
-
+        # Queue discipline
         self.__queue_discipline = ["FIFO", "LIFO", "PBS"]
-
+        # Verify if all variables are correct
         self.__verify_variables(a, s, c, k, n, d, pbs_distribution, pbs_parameters)
 
+        # Saving input variables
         self.__save_a = a
         self.__save_a_params = a_parameters
         self.__save_s = s
         self.__save_s_params = s_parameters
 
+        # Variables input for simualtion
         self.__a = self.__probability_distribution[self.__save_a](self.__save_a_params)
         self.__s = self.__probability_distribution[self.__save_s](self.__save_s_params)
         self.__c = c
@@ -63,6 +67,10 @@ class QueueingSimulation:
         self.__n = n
         self.__d = d
 
+        # This variables are assigned later
+        self.__simulation_time = 0
+
+        # PBS parameters
         self.__pbs_parameters = pbs_parameters
         self.__pbs_distribution = pbs_distribution
         if d == "PBS":
@@ -70,14 +78,25 @@ class QueueingSimulation:
                 self.__pbs_parameters
             )
 
+        # Simulation results
         self.result_simulation = pd.DataFrame()
         self.__simulation_time = 0
         self.number_probabilities = dict()
 
-    def __str__(self):
+    def __str__(self) -> str:
+        """Print dataset
+
+        Returns:
+            str: Dataframe in str mode
+        """
         return self.__result_simulation.to_string()
 
-    def _repr_html_(self):
+    def _repr_html_(self) -> pd.DataFrame:
+        """Print DataFrames in jupyter notebooks
+
+        Returns:
+            pd.DataFrame: Simulation result
+        """
         return self.__result_simulation._repr_html_()
 
     def __getitem__(self, key):
@@ -103,8 +122,8 @@ class QueueingSimulation:
             k (float): Maximum system capacity.
             n (float): Total population of potential customers.
             d (str): Queue discipline.
-            pbs_distribution (str): Label Distribution.
-            pbs_parameters (dict): Parameters of the PBS Distribution
+            pbs_distribution (str): Label Distribution. It can only be used if d='PBS' is selected
+            pbs_parameters (dict): Parameters of the PBS Distribution. It can only be used if d='PBS' is selected
         """
 
         # Verify if a and s belong to a actual probability distribution
@@ -131,30 +150,36 @@ class QueueingSimulation:
                 f"""'n' has to be a number and cannot be less or equals than zero."""
             )
 
+        # Review if the discipline is in the list
         if d not in self.__queue_discipline:
             raise ValueError(
                 f"""'d' has to be one of the following queue discipline: '{"', '".join(self.__queue_discipline)}'."""
             )
 
+        # Maximum number of people should be greater or equals to the number of servers
         if k < c:
             raise ValueError(f"""'k' cannot be less than the number of servers (c)""")
 
[…]
