+++ phitter/simulation/own_distribution/own_distribution.py
@@ -3,7 +3,12 @@ import numpy as np
 
 
 class OwnDistributions:
-    def __init__(self, parameters):
+    def __init__(self, parameters: dict):
+        """Creates the "OwnDistributions" Class
+
+        Args:
+            parameters (dict): Parameters of that distribution. all keys should be numbers greater or equals than zero. All values must sum up to 1.
+        """
         self.__parameters = parameters
 
         self.__first_verification()
@@ -16,7 +21,8 @@ class OwnDistributions:
 
         self.__error_detections()
 
-    def __first_verification(self):
+    def __first_verification(self) -> None:
+        """Verify if the keys are integers greater than zero, and verify if all values are floats."""
         # Quick verification
         if isinstance(self.__parameters, dict) == False:
             raise ValueError("You must pass a dictionary")
@@ -34,7 +40,8 @@ class OwnDistributions:
             if isinstance(self.__parameters[key], float) == False:
                 raise ValueError(f"""All keys must be floats.""")
 
-    def __error_detections(self):
+    def __error_detections(self) -> None:
+        """Identify the values that are greater than 1 or less than 0. Verify if accumulative probabilities are less or greater than 1. Must sum 1"""
 
         for key in self.__parameters.keys():
             if self.__parameters[key] <= 0 or self.__parameters[key] >= 1:
@@ -57,7 +64,15 @@ class OwnDistributions:
                 f"""All probabilities must be add up to 1, your probabilities sum a total of {last}"""
             )
 
-    def ppf(self, probability: int):
+    def ppf(self, probability: int) -> int:
+        """Assign a label according to a probability given by the created distribution
+
+        Args:
+            probability (int): Number between 0 and 1
+
+        Returns:
+            int: Returns label according to probability
+        """
         for label in self.__acummulative_parameters.keys():
             if probability <= self.__acummulative_parameters[label]:
                 return label
