### 2024-10-14 2445c03
Simulation fixing some issues and adding server utilization
