+++ phitter/simulation/queueing_simulation/queueing_simulation.py
@@ -32,10 +32,8 @@ class QueueingSimulation:
             k (float, optional): Maximum system capacity. This is the maximum number of customers that the system can accommodate at any given time, including both those in service and those waiting in the queue. It defines the limit beyond which arriving customers are either turned away or blocked from entering the system. Defaults to float("inf").
             n (float, optional): Total population of potential customers. This denotes the overall number of potential customers who might require service from the system. It can be finite or infinite and affects the arrival rates and the modeling of the system, especially in closed queueing networks. Defaults to float("inf").
             d (str, optional): Queue discipline. This describes the rule or policy that determines the order in which customers are served. Common disciplines include First-In-First-Out ("FIFO"), Last-In-First-Out ("LIFO"), priority-based service ("PBS"). The queue discipline impacts waiting times and the overall fairness of the system.. Defaults to "FIFO".
-            simulation_time (float, optional): This variable defines the total duration of the simulation. It sets the length of time over which the simulation will model the system's behavior. Defaults to float("inf")
-            number_of_simulations (int, optional): Number of simulations of the process. Can also be considered as the number of days or number of times you want to simulate your scenario. Defaults to 1.
             pbs_distribution (str | None, optional): Discrete distribution that identifies the label of the pbs, this parameter can only be used with "d='PBS'". Distributions that can be used: 'own_distribution', 'bernoulli', 'binomial', 'geometric', 'hypergeometric', 'logarithmic', 'negative_binomial', 'poisson'. Defaults to None.
-            pbs_parameters (dict | None, optional): Parameters of the discrete distribution that identifies the label of the pbs, this parameter can only be used with "d='PBS'". If it is 'own-distribution' add labels in the following way (example): {0: 0.5, 1: 0.3, 2: 0.2}. Where the "key" corresponds to the label and the "value" the probability whose total sum must add up to 1; "keys" with greater importances are the smallers and always have to be numeric keys. You can add as labels as you need.
+            pbs_parameters (dict | None, optional): Parameters of the discrete distribution that identifies the label of the pbs, this parameter can only be used with "d='PBS'". If it is 'own-distribution' add labels in the following way (example): {0: 0.5, 1: 0.3, 2: 0.2}. Where the "key" corresponds to the label and the "value" the probability whose total sum must add up to 1; "keys" with greater importances are the greaters and always have to be numeric keys. You can add as labels as you need.
         """
 
         # All phitter probability distributions
@@ -217,6 +215,8 @@ class QueueingSimulation:
         # Servers Information - Depends on the number in self.__c
         for server in range(1, self.__c + 1):
             simulation_results[f"Time busy server {server}"] = [0]
+        for server in range(1, self.__c + 1):
+            simulation_results[f"Server {server} attended this element?"] = [0]
 
         # Selections according to the queue discipline
         if self.__d == "FIFO":
@@ -274,9 +274,13 @@ class QueueingSimulation:
 
         # Determine all the arrival hours if the number of people do not exceed the maximum or the arriving time is less than the Max. Simulation Time
         while arriving_time < simulation_time and population < self.__n:
-            arrivals.append(self.__a.ppf(random.random()))
-            arriving_time += arrivals[-1]
-            population += 1
+            arr = self.__a.ppf(random.random())
+            if arriving_time + arr < simulation_time:
+                arrivals.append(arr)
+                arriving_time += arrivals[-1]
+                population += 1
+            else:
+                break
 
         # Start simulation for each arrival
         for arrival in arrivals:
@@ -356,13 +360,20 @@ class QueueingSimulation:
                 simulation_results[f"Time busy server {first_server_available}"].append(
                     simulation_results["Leave Time"][-1]
                 )
+                # This server was the enchanged of help the element
+                simulation_results[
[…]
