### 2024-10-15 603898a
simulation with tests and all in init
