+++ phitter/__init__.py
@@ -3,3 +3,4 @@ __version__ = "0.0.8"
 from .main import PHITTER
 from phitter import continuous
 from phitter import discrete
+from phitter import simulation
