+++ tests/pytest/test_queue_simulation_and_own.py
@@ -1,7 +1,9 @@
-import pandas
+import pytest
+import pandas as pd
 import phitter
 
 
+# Test para simulación con distribución propia y política PBS
 def test_own_distribution_and_pbs():
     parameters = {0: 0.5, 1: 0.3, 2: 0.2}
     simulation = phitter.simulation.QueueingSimulation(
@@ -14,23 +16,26 @@ def test_own_distribution_and_pbs():
         pbs_distribution="own_distribution",
         pbs_parameters=parameters,
     )
+    # Ejecutar la simulación con 2000 iteraciones
     simulation.run(2000)
+    # Verificar que la probabilidad de terminar después de un cierto tiempo es <= 1
     assert simulation.probability_to_finish_after_time() <= 1
 
 
+# Test para verificar que la política FIFO genera intervalos de confianza correctos
 def test_confidence_interval_fifo():
     simulation = phitter.simulation.QueueingSimulation(
         "exponential", {"lambda": 5}, "exponential", {"lambda": 20}, 3, d="FIFO"
     )
+    # Obtener los intervalos de confianza para las métricas con 10 réplicas
     a, b = simulation.confidence_interval_metrics(2000, replications=10)
-    assert (
-        type(a) == pandas.DataFrame
-        and len(a) > 0
-        and type(b) == pandas.DataFrame
-        and len(b) > 0
-    )
+
+    # Verificar que ambos resultados son DataFrames de pandas y no están vacíos
+    assert isinstance(a, pd.DataFrame) and len(a) > 0
+    assert isinstance(b, pd.DataFrame) and len(b) > 0
 
 
+# Test para verificar que la política LIFO genera métricas y probabilidades correctas
 def test_lifo_metrics():
     simulation = phitter.simulation.QueueingSimulation(
         "exponential",
@@ -38,11 +43,17 @@ def test_lifo_metrics():
         "exponential",
         {"lambda": 20},
         3,
-        n=100,
-        k=3,
-        d="LIFO",
+        n=100,  # Número de eventos
+        k=3,  # Capacidad del sistema
+        d="LIFO",  # Política LIFO
     )
+    # Ejecutar la simulación con 2000 iteraciones
     simulation.run(2000)
+
+    # Obtener el resumen de métricas y la probabilidad numérica
     metrics = simulation.metrics_summary()
     prob = simulation.number_probability_summary()
-    assert len(metrics) > 0 and len(prob) > 0
+
+    # Verificar que las métricas y probabilidades no están vacías
+    assert len(metrics) > 0
+    assert len(prob) > 0
