+++ SIMULATION.md
@@ -1,0 +1,43 @@
+<p align="center">
+    <picture>
+        <source media="(prefers-color-scheme: dark)" srcset="https://gist.githubusercontent.com/phitterio/66bc7f3674eac01ae646e30ba697a6d7/raw/e96dbba0eb26b20d35e608fefc3984bd87f0010b/DarkPhitterLogo.svg" width="350">
+        <source media="(prefers-color-scheme: light)" srcset="https://gist.githubusercontent.com/phitterio/170ce460d7e766545265772525edecf6/raw/71b4867c6e5683455cf1d68bea5bea7eda55ce7d/LightPhitterLogo.svg" width="350">
+        <img alt="phitter-dark-logo" src="https://gist.githubusercontent.com/phitterio/170ce460d7e766545265772525edecf6/raw/71b4867c6e5683455cf1d68bea5bea7eda55ce7d/LightPhitterLogo.svg" width="350">
+    </picture>
+</p>
+
+<p align="center">
+    <a href="https://pypi.org/project/phitter" target="_blank">
+        <img src="https://img.shields.io/pypi/dm/phitter.svg?color=blue" alt="Downloads">
+    </a>
+    <a href="https://pypi.org/project/phitter" target="_blank">
+        <img src="https://img.shields.io/badge/License-MIT-blue.svg" alt="License">
+    </a>
+    <a href="https://pypi.org/project/phitter" target="_blank">
+        <img src="https://img.shields.io/pypi/pyversions/phitter?color=blue" alt="Supported Python versions">
+    </a>
+    <a href="https://github.com/phitterio/phitter-kernel/actions/workflows/unittest.yml" target="_blank">
+        <img src="https://github.com/phitterio/phitter-kernel/actions/workflows/unittest.yml/badge.svg" alt="Tests">
+    </a>
+</p>
+
+<p>
+    Phitter analyzes datasets and determines the best analytical probability distributions that represent them. Phitter studies over 80 probability distributions, both continuous and discrete, 3 goodness-of-fit tests, and interactive visualizations. For each selected probability distribution, a standard modeling guide is provided along with spreadsheets that detail the methodology for using the chosen distribution in data science, operations research, and artificial intelligence.
+</p>
+<p>
+    This repository contains the implementation of the python library and the kernel of <a href="https://phitter.io">Phitter Web</a>
+</p>
+
+## Installation
+
+### Requirements
+
+```console
+python: >=3.9
+```
+
+### PyPI
+
+```console
+pip install phitter
+```
