### 2024-10-17 45841d2
Simulation updating general readme and simulation readme. Adding dependencies and adding comments to process simulation.🚀
