+++ README.md
@@ -23,6 +23,9 @@
 
 <p>
     Phitter analyzes datasets and determines the best analytical probability distributions that represent them. Phitter studies over 80 probability distributions, both continuous and discrete, 3 goodness-of-fit tests, and interactive visualizations. For each selected probability distribution, a standard modeling guide is provided along with spreadsheets that detail the methodology for using the chosen distribution in data science, operations research, and artificial intelligence.
+
+    In addition, Phitter offers the capability to perform process simulations, allowing users to graph and observe minimum times for specific observations. It also supports queue simulations with flexibility to configure various parameters, such as the number of servers, maximum population size, system capacity, and different queue disciplines, including First-In-First-Out (FIFO), Last-In-First-Out (LIFO), and priority-based service (PBS).
+
 </p>
 <p>
     This repository contains the implementation of the python library and the kernel of <a href="https://phitter.io">Phitter Web</a>
@@ -257,81 +260,81 @@ distribution.mode # -> 733.3333333333333
 
 #### 2. Resources Continuous Distributions
 
-| Distribution              | Phitter Playground                                                                                           | Excel File                                                                                                                         | Google Sheets Files                                                                                                   |
-| :------------------------ | :----------------------------------------------------------------------------------------------------------- | :--------------------------------------------------------------------------------------------------------------------------------- | :-------------------------------------------------------------------------------------------------------------------- |
-| alpha                     | ▶️[phitter:alpha](https://phitter.io/distributions/continuous/alpha)                                         | 📊[alpha.xlsx](https://github.com/phitterio/phitter-files/blob/main/continuous/alpha.xlsx)                                         | 🌐[gs:alpha](https://docs.google.com/spreadsheets/d/1yRovxx1YbqgEul65DjjXetysc_4qgX2a_2NQQA1AxCA)                     |
-| arcsine                   | ▶️[phitter:arcsine](https://phitter.io/distributions/continuous/arcsine)                                     | 📊[arcsine.xlsx](https://github.com/phitterio/phitter-files/blob/main/continuous/arcsine.xlsx)                                     | 🌐[gs:arcsine](https://docs.google.com/spreadsheets/d/1q8SKX4gmSbpGzimRvjopzaZ4KrEV5NY1EPmf1G1T7NQ)                   |
-| argus                     | ▶️[phitter:argus](https://phitter.io/distributions/continuous/argus)                                         | 📊[argus.xlsx](https://github.com/phitterio/phitter-files/blob/main/continuous/argus.xlsx)                                         | 🌐[gs:argus](https://docs.google.com/spreadsheets/d/1u2x7IFUSB7rEyhs7s6-C2btT1Bk5aCr4WiUYEML-8xs)                     |
[…]
