+++ pyproject.toml
@@ -36,7 +36,9 @@ dependencies = [
     "scipy>=1.1.0",
     "plotly>=5.14.0",
     "kaleido>=0.2.1",
-    "matplotlib>=3.3"
+    "matplotlib>=3.3",
+    "graphviz>=0.5.1",
+    "IPython>=0.13"
 ]
 
 [project.urls]

[…]
