### 2024-10-17 #7326 @kanishkan91
@ziyiyin97 would you be interested in reviewing this paper for JOSS? Seemed like a good fit. Let me know what you think.
