### 2024-10-17 #7326 @kanishkan91
@dannys4 would you be interested in reviewing this manuscript for JOSS?
