### 2024-10-18 b3ab8d5
simulation updating process simulation with df results and by stage and Readme updated
