+++ phitter/simulation/process_simulation/process_simulation.py
@@ -2,6 +2,7 @@ import math
 import random
 
 import numpy as np
+import pandas as pd
 from graphviz import Digraph
 from IPython.display import display
 
@@ -22,6 +23,31 @@ class ProcessSimulation:
             phitter.continuous.CONTINUOUS_DISTRIBUTIONS
             | phitter.discrete.DISCRETE_DISTRIBUTIONS
         )
+        self.servers = dict()
+
+        self.simulation_result = dict()
+
+    def __str__(self) -> str:
+        """Print dataset
+
+        Returns:
+            str: Dataframe in str mode
+        """
+        sim = pd.DataFrame(self.simulation_result)
+        return sim.to_string()
+
+    def _repr_html_(self) -> pd.DataFrame:
+        """Print DataFrames in jupyter notebooks
+
+        Returns:
+            pd.DataFrame: Simulation result
+        """
+        sim = pd.DataFrame(self.simulation_result)
+        return sim._repr_html_()
+
+    def __getitem__(self, key):
+        sim = pd.DataFrame(self.simulation_result)
+        return sim[key]
 
     def add_process(
         self,
@@ -29,6 +55,7 @@ class ProcessSimulation:
         parameters: dict,
         process_id: str,
         number_of_products: int = 1,
+        number_of_servers: int = 1,
         new_branch: bool = False,
         previous_ids: list[str] = None,
     ) -> None:
@@ -38,7 +65,8 @@ class ProcessSimulation:
             prob_distribution (str): Probability distribution to be used. You can use one of the following: 'alpha', 'arcsine', 'argus', 'beta', 'beta_prime', 'beta_prime_4p', 'bradford', 'burr', 'burr_4p', 'cauchy', 'chi_square', 'chi_square_3p', 'dagum', 'dagum_4p', 'erlang', 'erlang_3p', 'error_function', 'exponential', 'exponential_2p', 'f', 'fatigue_life', 'folded_normal', 'frechet', 'f_4p', 'gamma', 'gamma_3p', 'generalized_extreme_value', 'generalized_gamma', 'generalized_gamma_4p', 'generalized_logistic', 'generalized_normal', 'generalized_pareto', 'gibrat', 'gumbel_left', 'gumbel_right', 'half_normal', 'hyperbolic_secant', 'inverse_gamma', 'inverse_gamma_3p', 'inverse_gaussian', 'inverse_gaussian_3p', 'johnson_sb', 'johnson_su', 'kumaraswamy', 'laplace', 'levy', 'loggamma', 'logistic', 'loglogistic', 'loglogistic_3p', 'lognormal', 'maxwell', 'moyal', 'nakagami', 'non_central_chi_square', 'non_central_f', 'non_central_t_student', 'normal', 'pareto_first_kind', 'pareto_second_kind', 'pert', 'power_function', 'rayleigh', 'reciprocal', 'rice', 'semicircular', 'trapezoidal', 'triangular', 't_student', 't_student_3p', 'uniform', 'weibull', 'weibull_3p', 'bernoulli', 'binomial', 'geometric', 'hypergeometric', 'logarithmic', 'negative_binomial', 'poisson'.
             parameters (dict): Parameters of the probability distribution.
             process_id (str):  Unique name of the process to be simulated
-            number_of_products (int ,optional): Number of elements that are need to simulate in that stage. Value has to be greater than 0. Defaults equals to 1.
+            number_of_products (int ,optional): Number of elements that are needed to simulate in that stage. Value has to be greater than 0. Defaults equals to 1.
+            number_of_servers (int, optional): Number of servers that process has and are needed to simulate in that stage. Value has to be greater than 0. Defaults equals to 1.
[…]
