### 2024-10-23 #7326 @dannys4
I'd be happy to provide a review, but I'm not going to have much insight into the code or dependencies in the context of the wider Python ecosystem; I can write competent Python, it's just not my forte. The website states "Phitter is not fast, it is exhaustive and free", which suggests this is fine (though the article espouses its efficiency). Seems interesting though!
