### 2024-10-29 667bd1f
Write readme module fit and simulation
