+++ phitter/simulation/process_simulation/process_simulation.py
@@ -1,7 +1,6 @@
 import math
 import random
 
-import numpy as np
 import pandas as pd
 from graphviz import Digraph
 from IPython.display import display
@@ -19,10 +18,7 @@ class ProcessSimulation:
         self.number_of_products = dict()
         self.process_positions = dict()
         self.next_process = dict()
-        self.probability_distribution = (
-            phitter.continuous.CONTINUOUS_DISTRIBUTIONS
-            | phitter.discrete.DISCRETE_DISTRIBUTIONS
-        )
+        self.probability_distribution = phitter.continuous.CONTINUOUS_DISTRIBUTIONS | phitter.discrete.DISCRETE_DISTRIBUTIONS
         self.servers = dict()
 
         self.simulation_result = dict()
@@ -72,9 +68,7 @@ class ProcessSimulation:
         """
         # Verify if the probability is created in phitter
         if prob_distribution not in self.probability_distribution.keys():
-            raise ValueError(
-                f"""Please select one of the following probability distributions: '{"', '".join(self.probability_distribution.keys())}'."""
-            )
+            raise ValueError(f"""Please select one of the following probability distributions: '{"', '".join(self.probability_distribution.keys())}'.""")
         else:
             # Verify unique id name for each process
             if process_id not in self.order.keys():
@@ -84,9 +78,7 @@ class ProcessSimulation:
                     if number_of_servers >= 1:
                         # Verify that if you create a new branch, it's impossible to have a previous id (or preceding process). One of those is incorrect
                         if new_branch == True and previous_ids != None:
-                            raise ValueError(
-                                f"""You cannot select 'new_branch' is equals to True if 'previous_id' is not empty. OR you cannot add 'previous_ids' if 'new_branch' is equals to True."""
-                            )
+                            raise ValueError(f"""You cannot select 'new_branch' is equals to True if 'previous_id' is not empty. OR you cannot add 'previous_ids' if 'new_branch' is equals to True.""")
                         else:
                             # If it is a new branch then initialize all the needed paramters
                             if new_branch == True:
@@ -95,27 +87,17 @@ class ProcessSimulation:
                                 self.order[process_id] = branch_id
                                 self.number_of_products[process_id] = number_of_products
                                 self.servers[process_id] = number_of_servers
-                                self.process_prob_distr[process_id] = (
-                                    self.probability_distribution[prob_distribution](
-                                        parameters
-                                    )
-                                )
+                                self.process_prob_distr[process_id] = self.probability_distribution[prob_distribution](parameters)
                                 self.next_process[process_id] = 0
                                 # Create id of that process in the simulation result
                                 self.simulation_result[process_id] = []
 
                             # If it is NOT a new branch then initialize all the needed paramters
-                            elif previous_ids != None and all(
-                                id in self.order.keys() for id in previous_ids
-                            ):
+                            elif previous_ids != None and all(id in self.order.keys() for id in previous_ids):
                                 self.order[process_id] = previous_ids
                                 self.number_of_products[process_id] = number_of_products
                                 self.servers[process_id] = number_of_servers
-                                self.process_prob_distr[process_id] = (
-                                    self.probability_distribution[prob_distribution](
-                                        parameters
-                                    )
-                                )
+                                self.process_prob_distr[process_id] = self.probability_distribution[prob_distribution](parameters)
                                 self.next_process[process_id] = 0
                                 # Create id of that process in the simulation result
                                 self.simulation_result[process_id] = []
@@ -127,17 +109,11 @@ class ProcessSimulation:
[…]
