+++ README.md
@@ -23,10 +23,12 @@
 
 <p>
     Phitter analyzes datasets and determines the best analytical probability distributions that represent them. Phitter studies over 80 probability distributions, both continuous and discrete, 3 goodness-of-fit tests, and interactive visualizations. For each selected probability distribution, a standard modeling guide is provided along with spreadsheets that detail the methodology for using the chosen distribution in data science, operations research, and artificial intelligence.
-
-    In addition, Phitter offers the capability to perform process simulations, allowing users to graph and observe minimum times for specific observations. It also supports queue simulations with flexibility to configure various parameters, such as the number of servers, maximum population size, system capacity, and different queue disciplines, including First-In-First-Out (FIFO), Last-In-First-Out (LIFO), and priority-based service (PBS).
-
 </p>
+
+<p>
+    In addition, Phitter offers the capability to perform process simulations, allowing users to graph and observe minimum times for specific observations. It also supports queue simulations with flexibility to configure various parameters, such as the number of servers, maximum population size, system capacity, and different queue disciplines, including First-In-First-Out (FIFO), Last-In-First-Out (LIFO), and priority-based service (PBS).
+</p>
+
 <p>
     This repository contains the implementation of the python library and the kernel of <a href="https://phitter.io">Phitter Web</a>
 </p>
@@ -45,9 +47,11 @@ python: >=3.9
 pip install phitter
 ```
 
+
+
 ## Usage
 
-### Notebook's Tutorials
+### ***1. Fit Notebook's Tutorials***
 
 |             Tutorial             |                                                                                                                  Notebooks                                                                                                                   |
 | :------------------------------: | :------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------: |
@@ -57,7 +61,28 @@ pip install phitter
 |   **Fit Specific Disribution**   | <a target="_blank" href="https://colab.research.google.com/github/phitterio/phitter-kernel/blob/main/examples/fit_specefic_distribution.ipynb"><img src="https://colab.research.google.com/assets/colab-badge.svg" alt="Open In Colab"/></a> |
 |     **Working Distribution**     |   <a target="_blank" href="https://colab.research.google.com/github/phitterio/phitter-kernel/blob/main/examples/working_distribution.ipynb"><img src="https://colab.research.google.com/assets/colab-badge.svg" alt="Open In Colab"/></a>    |
 
-### General
+### ***2. Simulation Notebook's Tutorials***
+pending
+
+## Documentation
+
+
+
+
+
+
+
+
+
+
+
+<details>
+
+<summary style="font-size: 16px; font-weight: bold;">Documentation Fit Module</summary>
+
+
+
+### General Fit
 
 ```python
 import phitter
@@ -465,6 +490,233 @@ distribution.mode # -> 733.3333333333333
 |     negative_binomial      | 0.0293 | 0.0000 | 0.0000 | 0.0000 | 0.0000 | 0.0000 |
 |          poisson           | 0.0000 | 0.0000 | 0.0000 | 0.0000 | 0.0000 | 0.0000 |
 |          uniform           | 0.0000 | 0.0000 | 0.0000 | 0.0000 | 0.0000 | 0.0000 |
+
+</details>
+
+
+
+
+
+
+
+
+
+<details>
+<summary style="font-size: 16px; font-weight: bold;">Documentation Simulation Module</summary>
+
+
+
+## Process Simulation
+
+This will help you to understand your processes. To use it, run the following line
+
+```python
+from phitter import simulation
+
+# Create a simulation process instance
[…]

[…]
