### 2024-10-29 749a0d0
include IPython requirements
