+++ requirements.txt
@@ -2,7 +2,7 @@ scipy>=1.1.0
 plotly>=5.14.0
 kaleido>=0.2.1
 matplotlib>=3.3
-
+IPython>=0.13
 pytest
 pandas>=1.5.0
 graphviz>=0.5.1
