+++ pyproject.toml
@@ -39,7 +39,8 @@ dependencies = [
     "kaleido>=0.2.1",
     "matplotlib>=3.3",
     "pandas>=1.5.0",
-    "graphviz>=0.5.1"
+    "graphviz>=0.5.1",
+    "IPython>=0.13"
 ]
 
 [project.urls]
