+++ phitter/simulation/own_distribution/own_distribution.py
@@ -1,8 +1,3 @@
-import random
-
-import numpy as np
-
-
 class OwnDistributions:
     def __init__(self, parameters: dict):
         """Creates the "OwnDistributions" Class
@@ -30,13 +25,9 @@ class OwnDistributions:
 
         for key in self.__parameters.keys():
             if isinstance(key, int) == False:
-                raise ValueError(
-                    f"""All keys must be integers greater or equal than 0."""
-                )
+                raise ValueError(f"""All keys must be integers greater or equal than 0.""")
             elif key < 0:
-                raise ValueError(
-                    f"""All keys must be integers greater or equal than 0."""
-                )
+                raise ValueError(f"""All keys must be integers greater or equal than 0.""")
 
             if isinstance(self.__parameters[key], float) == False:
                 raise ValueError(f"""All keys must be floats.""")
@@ -46,24 +37,15 @@ class OwnDistributions:
 
         for key in self.__parameters.keys():
             if self.__parameters[key] <= 0 or self.__parameters[key] >= 1:
-                raise ValueError(
-                    f"""All probabilities must be greater than 0 and less than 1. You have a value of {self.__parameters[key]} for key {key}"""
-                )
+                raise ValueError(f"""All probabilities must be greater than 0 and less than 1. You have a value of {self.__parameters[key]} for key {key}""")
 
-            if (
-                self.__acummulative_parameters[key] > 1
-                or self.__acummulative_parameters[key] <= 0
-            ):
-                raise ValueError(
-                    f"""All probabilities must be add up to 1 and must be greater than 0. You have a acummulative value of {self.__acummulative_parameters[key]}"""
-                )
+            if self.__acummulative_parameters[key] > 1 or self.__acummulative_parameters[key] <= 0:
+                raise ValueError(f"""All probabilities must be add up to 1 and must be greater than 0. You have a acummulative value of {self.__acummulative_parameters[key]}""")
             else:
                 last = self.__acummulative_parameters[key]
 
         if last != 1:
-            raise ValueError(
-                f"""All probabilities must be add up to 1, your probabilities sum a total of {last}"""
-            )
+            raise ValueError(f"""All probabilities must be add up to 1, your probabilities sum a total of {last}""")
 
     def ppf(self, probability: int) -> int:
         """Assign a label according to a probability given by the created distribution
