+++ phitter/simulation/queueing_simulation/queueing_simulation.py
@@ -39,15 +39,10 @@ class QueueingSimulation:
         """
 
         # All phitter probability distributions
-        self.__probability_distribution = (
-            phitter.continuous.CONTINUOUS_DISTRIBUTIONS
-            | phitter.discrete.DISCRETE_DISTRIBUTIONS
-        )
+        self.__probability_distribution = phitter.continuous.CONTINUOUS_DISTRIBUTIONS | phitter.discrete.DISCRETE_DISTRIBUTIONS
 
         # Distributions for labels
-        self.__pbs_distributions = {
-            "own_distribution": OwnDistributions
-        } | phitter.discrete.DISCRETE_DISTRIBUTIONS
+        self.__pbs_distributions = {"own_distribution": OwnDistributions} | phitter.discrete.DISCRETE_DISTRIBUTIONS
         # Queue discipline
         self.__queue_discipline = ["FIFO", "LIFO", "PBS"]
         # Verify if all variables are correct
@@ -74,9 +69,7 @@ class QueueingSimulation:
         self.__pbs_parameters = pbs_parameters
         self.__pbs_distribution = pbs_distribution
         if d == "PBS":
-            self.__label = self.__pbs_distributions[self.__pbs_distribution](
-                self.__pbs_parameters
-            )
+            self.__label = self.__pbs_distributions[self.__pbs_distribution](self.__pbs_parameters)
 
         # Simulation results
         self.result_simulation = pd.DataFrame()
@@ -127,34 +120,21 @@ class QueueingSimulation:
         """
 
         # Verify if a and s belong to a actual probability distribution
-        if (
-            a not in self.__probability_distribution.keys()
-            or s not in self.__probability_distribution.keys()
-        ):
-            raise ValueError(
-                f"""Please select one of the following probability distributions: '{"', '".join(self.__probability_distribution.keys())}'."""
-            )
+        if a not in self.__probability_distribution.keys() or s not in self.__probability_distribution.keys():
+            raise ValueError(f"""Please select one of the following probability distributions: '{"', '".join(self.__probability_distribution.keys())}'.""")
         # Verify Number of Servers
         if c <= 0:
-            raise ValueError(
-                f"""'c' has to be a number and cannot be less or equals than zero."""
-            )
+            raise ValueError(f"""'c' has to be a number and cannot be less or equals than zero.""")
         # Verify Maximum System Capacity
         if k <= 0:
-            raise ValueError(
-                f"""'k' has to be a number and cannot be less or equals than zero."""
-            )
+            raise ValueError(f"""'k' has to be a number and cannot be less or equals than zero.""")
         # Verify Total population of potential customers.
         if n <= 0:
-            raise ValueError(
-                f"""'n' has to be a number and cannot be less or equals than zero."""
-            )
+            raise ValueError(f"""'n' has to be a number and cannot be less or equals than zero.""")
 
         # Review if the discipline is in the list
         if d not in self.__queue_discipline:
-            raise ValueError(
-                f"""'d' has to be one of the following queue discipline: '{"', '".join(self.__queue_discipline)}'."""
-            )
+            raise ValueError(f"""'d' has to be one of the following queue discipline: '{"', '".join(self.__queue_discipline)}'.""")
 
         # Maximum number of people should be greater or equals to the number of servers
         if k < c:
@@ -165,19 +145,13 @@ class QueueingSimulation:
             if pbs_distribution != None and pbs_parameters != None:
                 # Review if the selected distribution was created
                 if pbs_distribution not in self.__pbs_distributions:
-                    raise ValueError(
-                        f"""You should select one of the following distributions: {self.__pbs_distributions}"""
-                    )
[…]
