+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "phitter"
 version = "1.0.0"
-description = "Find the best probability distribution for your dataset"
+description = "Find the best probability distribution for your dataset and simulate processes and queues"
 authors = [{name = "Sebastián José Herrera Monterrosa",  email = "phitter.email@gmail.com"}, {name = "Carlos Andrés Másmela Pinilla",  email = "phitter.email@gmail.com"}]
 readme = "README.md"
 requires-python = ">=3.9"
@@ -49,4 +49,4 @@ Documentation = "https://github.com/phitterio/phitter-kernel"
 Repository = "https://github.com/phitterio/phitter-kernel"
 
 [tool.setuptools]
-packages = ["phitter", "phitter.continuous", "phitter.continuous.continuous_distributions", "phitter.continuous.continuous_measures", "phitter.continuous.continuous_statistical_tests", "phitter.discrete", "phitter.discrete.discrete_distributions", "phitter.discrete.discrete_measures", "phitter.discrete.discrete_statistical_tests"]
+packages = ["phitter", "phitter.continuous", "phitter.continuous.continuous_distributions", "phitter.continuous.continuous_measures", "phitter.continuous.continuous_statistical_tests", "phitter.discrete", "phitter.discrete.discrete_distributions", "phitter.discrete.discrete_measures", "phitter.discrete.discrete_statistical_tests", "phitter.simulation", "phitter.simulation.own_distributions", "phitter.simulation.process_simulation", "phitter.simulation.queueing_simulation"]
