### 2024-11-06 21ae16c
Github actions: run tests when changes phitter folder
