### 2025-01-03 #7326 @kanishkan91
@baxmittens based on your previous publication in JOSS, would you be interested in reviewing this paper?
