### 2025-01-03 #7326 @kanishkan91
@EwoutH would you be interested in reviewing this paper for JOSS?
