### 2025-01-03 #7625 @editorialbot
[REVIEW]: Phitter
**Submitting author:** @sebastianherreramonterrosa (<a href="http://orcid.org/0009-0002-2766-642X">Sebastián Herrera</a>)
**Repository:** https://github.com/phitterio/phitter-kernel
**Branch with paper.md** (empty if default branch): main
**Version:** v0.7.1
