### 2025-01-03 #7625 @kanishkan91
@EwoutH- I've added some general instructions above. Your JOSS review is guided by a checklist which you can generate. Feel free to reach out to me in case you have any specific questions by pinging me directly in this thread.
