### 2025-01-03 phitterio/phitter-kernel#3 @mdhaber
Typos
While reviewing the JOSS submission (openjournals/joss-reviews#7625), I'll note typos and such here:

- "Distribution" is missing the first "t" in many places
- Link to "Fit Specific Distribution" is broken due to misspelling of "specific"

![image](https://github.com/user-attachments/assets/dd9d044c-113f-4008-b990-64034bed81c6)
