### 2025-01-03 #7625 @EwoutH
Really cool project @sebastianherreramonterrosa! I definitely see myself using this.

Open questions/suggestions:

1. Contribution and authorship: It seems Carlos Másmela has made significant contributions to the repo. Should he be listed as a co-author?
2. Functionality
    a. The paper claims "over 80 probability distributions". I think I found them here: https://github.com/phitterio/phitter-kernel/blob/main/documentation/FIT.md#continuous-distributions, but it might be nice to hyperlink to the documentation in the paper so users can find this list.
    b. "Accelerated fitting capabilities for large datasets (over 100K samples)" --> Where can I find an example of this?
3. Documentation
    a. Installation instructions --> Maybe link to docs here, especially "Is there a clearly-stated list of dependencies?" seems to be a JOSS requirement (@kanishkan91 does that need to be in the main paper, or can that be linked to in docs?)
    b. Example usage is not present. A few simple examples (like in the Readme or docs) might be useful.
    c. Functionality documentation --> Idem dito (@kanishkan91 I thought JOSS was [against](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) API documentation? What's the balance here?
        _Given this format, a “full length” paper is not permitted, and software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation._
    d. Community guidelines --> @kanishkan91 is [this](https://github.com/phitterio/phitter-kernel?tab=readme-ov-file#contribution) enough or does JOSS require more?
5. Software paper
    a. A statement of need / State of the field --> Might be nice to name the largest competitors of Phitter
    b. References --> There are currently none, it might be nice to refer to statistical foundations, related Python statistical libraries, other comparative software and applications.

Hope this helps!
