### 2025-01-03 phitterio/phitter-kernel#4 @mdhaber
Incorrect variance of normal distribution
Incorrect variance of
```python3
import phitter
phitter.continuous.NORMAL({"mu": 0, "sigma": 1}).variance  # 3; should be 1
```

I haven't looked into whether the problem is due to an mistake in a hard-coded value or an implementation that might affect other distributions.

Crossref openjournals/joss-reviews#7625
