### 2025-01-03 #7625 @mdhaber
I second most of @EwoutH's comments, although re: documentation, I think these documentation requirements are to be satisfied by the project documention, not the JOSS paper. I think there are plenty of examples at the GitHub repository, and the software dependencies are clear. I would suggest that the example code in the paper can be removed, but I think @kanishkan91 should be the judge of that per JOSS standards.

Other comments to follow.
