### 2025-01-05 6a919e5
1. Correction normal variance 2. Typos correction
