### 2025-01-05 phitterio/phitter-kernel#3 @sebastianherreramonterrosa
@mdhaber thanks for the heads up. The issue has been corrected.
