### 2025-01-15 #7625 @sebastianherreramonterrosa
Yes, I am still working on the project and the article. Here are some considerations and responses to the issues raised @EwoutH, @mdhaber.

- The scope of the Phitter project and its functionalities has been extended from the probability fitting module for datasets to include the process simulation module. This raises the question of whether the article needs to be modified due to this change.

- _Has the submitting author (@sebastianherreramonterrosa) made major contributions to the software? Does the full list of paper authors seem appropriate and complete? Yes, sebastianherreramonterrosa is the primary author. Should candemas97 be credited?_
  A: @candemas97 has been the developer of the simulation module, so they should appear as an author in the article.

- _Does the paper intend to present any original "data" or original "results"? I don't think so; I just want to make sure I'm not missing anything. See "Data sharing" and "Reproducibility" items of the reviewer checklists._
  A: Phitter does not aim to present any novel results.

- _phitter.io makes claims that are a little too strong for me to check the "Functionality" box without reservation:
  "Phitter analyzes datasets and determines the best analytical probability distributions that represent them." I'd suggest that it helps the user to find appropriate candidates; it does not itself "determine the best... probability distribution".
  "Phitter is not fast, it is exhaustive and free." While 80 distributions and three tests is substantial, "exhaustive" may be an overstatement. There are 120+ continuous probability distributions and half a dozen distinct general goodness-of-fit tests in SciPy, but I wouldn't consider it to be "exhaustive". Please consider a different adjective._
  A: Regarding the first item, suggesting appropriate candidates is better. As for the second, I was referring to the process, although this message should be revised. Here, I have a question: should the messages displayed in the application [phitter.io](https://phitter.io/) be as rigorous as those in the article, or can they differ?
