### 2025-01-15 #7625 @mdhaber
> Here, I have a question: should the messages displayed in the application phitter.io be as rigorous as those in the article, or can they differ?

I understand that the messages displaying phitter.io are intended to advertise the capabilities, and for that purpose, I don't think they need to be as _rigorous_ or _precisie_ as in the article, but I have to be able to check the box "Have the functional claims of the software been confirmed?" If I see claims on the website that are much stronger than those I can verify, I can't check the box.

> This raises the question of whether the article needs to be modified due to this change.

I think that depends on your response to our requests for references to existing work. JOSS's recommendations about "Substantial scholarly effort" are:

> JOSS publishes articles about software that represent substantial scholarly effort on the part of the authors. Your software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)... As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual.

I think the rule of thumb is satisfied, and the length of the commit history, number of commits, and total lines of code considtions are met. But I'm not sure whether it is "significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)." There are many fitting packages that appear similar. How is Phitter different from [`fitter`](https://github.com/cokelaer/fitter), for example? If the fitting functionality doesn't add much to the ecosystem, the paper might need to describe other features that are new/better that what already exists.
