### 2025-01-15 #7625 @sebastianherreramonterrosa
The main reasons for reimplementing each distribution are as follows:

1. Phitter aims to implement a faster method for calculating distribution parameters whenever possible. To achieve this, it attempts to solve the distribution's parametric equation system, which is significantly faster than the `.fit` method for all Scipy distributions. However, this objective is not achievable for every distribution. This table specifies the estimation times for each distribution based on sample size: [Estimation time parameters continuous distributions](https://github.com/phitterio/phitter-kernel/blob/main/documentation/FIT.md#estimation-time-parameters-continuous-distributions). For distributions where the estimation time is 0 for a given sample size, the parametric equation system is solved. In cases where the estimation time increases, Scipy's maximum likelihood estimation method is used. This approach contributes significantly to performance improvements.

2. Phitter adheres to the standard notation of each distribution, including its parameters and the implementation of corresponding methods. For every distribution, there is an Excel document and a Google Sheets document that define its characteristics. These documents contain the parametric formulas mentioned above and provide a structured workflow for those working with spreadsheets: [Resources Continuous Distributions](https://github.com/phitterio/phitter-kernel/blob/main/documentation/FIT.md#2-resources-continuous-distributions). For example, in the Rice distribution, the `pdf` method using Scipy is implemented as `scipy.stats.rice.pdf(x, self.v / self.sigma, scale=self.sigma)`, rather than as some users might expect: `scipy.stats.rice.pdf(x, self.v, scale=self.sigma)`. In conclusion, Phitter's development process involved first creating a spreadsheet for each distribution using the most widely recognized standard notation and then constructing the distribution using Scipy and Numpy.

3. As far as I know, Scipy does not provide a general implementation for the Anderson-Darling critical value. In Phitter, this is implemented following the proposal in this article: [Evaluating the Anderson-Darling Distribution](https://www.jstatsoft.org/article/view/v009i02).

---

I am unsure whether these considerations should be explicitly mentioned in the article or if they are sufficient to address the specified item. Additionally, the simulation module should also be included.
