### 2025-01-22 #7625 @mdhaber
OK, thanks for addressing how Phitter is intended to be faster / more standard than SciPy. I have some thoughts about those points, but I think the main thing that remains to decide whether the process simulation work should be featured in the paper is how Phitter improves upon comparable tools like `fitter` and `zfit`, which I think have similar goals. (SciPy is a much more general library.)

---

1. When you say "the parametric equation system is solved", do you mean you have implemented an *analytical* solution to the [*likelihood equations*](https://en.wikipedia.org/wiki/Maximum_likelihood_estimation), or something else? (Note that SciPy implements many of these analytical solutions already; see scipy/scipy#11782). I think I partially answered my own question: [chi_square_3p.get_parameters](https://github.com/phitterio/phitter-kernel/blob/22e6f4cfc378085b21b4a89106064afb2be906cb/phitter/continuous/continuous_distributions/chi_square_3p.py#L159) seems to implement something like method of moments for the 3-parameter chi-squared distribution, so that is not maximum likelihood estimation.) So it sounds like you have chosen a fitting method (e.g. maximum likelihood, method of moments, maximum product spacing) for each distribution depending on which is fastest?
2. You wrote "For every distribution, there is an Excel document and a Google Sheets document that define its characteristics." One of my questions above was about the references for this material. I'd suggest citing the sources you used to decide on the "most widely recognized standard notation" as well as the source of equations.
3. [`scipy.stats.goodness_of_fit` ](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.goodness_of_fit.html) offers Anderson-Darling p-values for any distribution and sample size using a Monte Carlo approach.
