### 2025-02-18 #7625 @mdhaber
Re: 2, I see the material, the comment was about _citing sources_ of the information.

>  I'd suggest citing the sources you used to decide on the "most widely recognized standard notation" as well as the source of equations.

I see the information listed in "Resources Continuous Distributions" and on the website; I don't see the source of that information.
