+++ phitter/continuous/continuous_distributions/loggamma.py
@@ -4,10 +4,10 @@ import scipy.special
 import scipy.stats
 
 
-class LOGGAMMA:
+class LogGamma:
     """
-    Loggamma distribution
-    Parameters LOGGAMMA distribution: {"c": *, "mu": *, "sigma": *}
+    LogGamma distribution
+    Parameters LogGamma Distribution: {"c": *, "mu": *, "sigma": *}
     https://phitter.io/distributions/continuous/loggamma
     """
 
@@ -18,13 +18,13 @@ class LOGGAMMA:
         init_parameters_examples=False,
     ):
         """
-        Initializes the LOGGAMMA distribution by either providing a Continuous Measures instance [CONTINUOUS_MEASURES] or a dictionary with the distribution's parameters.
-        Parameters LOGGAMMA distribution: {"c": *, "mu": *, "sigma": *}
+        Initializes the LogGamma Distribution by either providing a Continuous Measures instance [ContinuousMeasures] or a dictionary with the distribution's parameters.
+        Parameters LogGamma Distribution: {"c": *, "mu": *, "sigma": *}
         https://phitter.io/distributions/continuous/loggamma
         """
         if continuous_measures is None and parameters is None and init_parameters_examples == False:
             raise ValueError(
-                "You must initialize the distribution by providing one of the following: distribution parameters, a Continuous Measures [CONTINUOUS_MEASURES] instance, or by setting init_parameters_examples to True."
+                "You must initialize the distribution by providing one of the following: distribution parameters, a Continuous Measures [ContinuousMeasures] instance, or by setting init_parameters_examples to True."
             )
         if continuous_measures != None:
             self.parameters = self.get_parameters(continuous_measures=continuous_measures)
@@ -201,7 +201,7 @@ if __name__ == "__main__":
     import sys
 
     sys.path.append("../")
-    from continuous_measures import CONTINUOUS_MEASURES
+    from continuous_measures import ContinuousMeasures
 
     ## Import function to get continuous_measures
     def get_data(path: str) -> list[float]:
@@ -213,8 +213,8 @@ if __name__ == "__main__":
     ## Distribution class
     path = "../continuous_distributions_sample/sample_loggamma.txt"
     data = get_data(path)
-    continuous_measures = CONTINUOUS_MEASURES(data)
-    distribution = LOGGAMMA(continuous_measures=continuous_measures)
+    continuous_measures = ContinuousMeasures(data)
+    distribution = LogGamma(continuous_measures=continuous_measures)
 
     print(f"{distribution.name} distribution")
     print(f"Parameters: {distribution.parameters}")
