+++ phitter/continuous/continuous_distributions/inverse_gamma.py
@@ -4,11 +4,11 @@ import scipy.special
 import scipy.stats
 
 
-class INVERSE_GAMMA:
+class InverseGamma:
     """
-    Inverse Gamma distribution
+    Inverse Gamma Distribution
     Also known Pearson Type 5 distribution
-    Parameters INVERSE_GAMMA distribution: {"alpha": *, "beta": *}
+    Parameters InverseGamma Distribution: {"alpha": *, "beta": *}
     https://phitter.io/distributions/continuous/inverse_gamma
     """
 
@@ -19,13 +19,13 @@ class INVERSE_GAMMA:
         init_parameters_examples=False,
     ):
         """
-        Initializes the INVERSE_GAMMA distribution by either providing a Continuous Measures instance [CONTINUOUS_MEASURES] or a dictionary with the distribution's parameters.
-        Parameters INVERSE_GAMMA distribution: {"alpha": *, "beta": *}
+        Initializes the InverseGamma Distribution by either providing a Continuous Measures instance [ContinuousMeasures] or a dictionary with the distribution's parameters.
+        Parameters InverseGamma Distribution: {"alpha": *, "beta": *}
         https://phitter.io/distributions/continuous/inverse_gamma
         """
         if continuous_measures is None and parameters is None and init_parameters_examples == False:
             raise ValueError(
-                "You must initialize the distribution by providing one of the following: distribution parameters, a Continuous Measures [CONTINUOUS_MEASURES] instance, or by setting init_parameters_examples to True."
+                "You must initialize the distribution by providing one of the following: distribution parameters, a Continuous Measures [ContinuousMeasures] instance, or by setting init_parameters_examples to True."
             )
         if continuous_measures != None:
             self.parameters = self.get_parameters(continuous_measures=continuous_measures)
@@ -241,7 +241,7 @@ if __name__ == "__main__":
     import sys
 
     sys.path.append("../")
-    from continuous_measures import CONTINUOUS_MEASURES
+    from continuous_measures import ContinuousMeasures
 
     ## Import function to get continuous_measures
     def get_data(path: str) -> list[float]:
@@ -253,8 +253,8 @@ if __name__ == "__main__":
     ## Distribution class
     path = "../continuous_distributions_sample/sample_inverse_gamma.txt"
     data = get_data(path)
-    continuous_measures = CONTINUOUS_MEASURES(data)
-    distribution = INVERSE_GAMMA(continuous_measures=continuous_measures)
+    continuous_measures = ContinuousMeasures(data)
+    distribution = InverseGamma(continuous_measures=continuous_measures)
 
     print(f"{distribution.name} distribution")
     print(f"Parameters: {distribution.parameters}")
