+++ phitter/continuous/continuous_distributions/gamma.py
@@ -3,10 +3,10 @@ import scipy.special
 import scipy.stats
 
 
-class GAMMA:
+class Gamma:
     """
     Gamma distribution
-    Parameters GAMMA distribution: {"alpha": *, "beta": *}
+    Parameters Gamma Distribution: {"alpha": *, "beta": *}
     https://phitter.io/distributions/continuous/gamma
     """
 
@@ -17,13 +17,13 @@ class GAMMA:
         init_parameters_examples=False,
     ):
         """
-        Initializes the GAMMA distribution by either providing a Continuous Measures instance [CONTINUOUS_MEASURES] or a dictionary with the distribution's parameters.
-        Parameters GAMMA distribution: {"alpha": *, "beta": *}
+        Initializes the Gamma Distribution by either providing a Continuous Measures instance [ContinuousMeasures] or a dictionary with the distribution's parameters.
+        Parameters Gamma Distribution: {"alpha": *, "beta": *}
         https://phitter.io/distributions/continuous/gamma
         """
         if continuous_measures is None and parameters is None and init_parameters_examples == False:
             raise ValueError(
-                "You must initialize the distribution by providing one of the following: distribution parameters, a Continuous Measures [CONTINUOUS_MEASURES] instance, or by setting init_parameters_examples to True."
+                "You must initialize the distribution by providing one of the following: distribution parameters, a Continuous Measures [ContinuousMeasures] instance, or by setting init_parameters_examples to True."
             )
         if continuous_measures != None:
             self.parameters = self.get_parameters(continuous_measures=continuous_measures)
@@ -197,7 +197,7 @@ if __name__ == "__main__":
     import sys
 
     sys.path.append("../")
-    from continuous_measures import CONTINUOUS_MEASURES
+    from continuous_measures import ContinuousMeasures
 
     ## Import function to get continuous_measures
     def get_data(path: str) -> list[float]:
@@ -209,8 +209,8 @@ if __name__ == "__main__":
     ## Distribution class
     path = "../continuous_distributions_sample/sample_gamma.txt"
     data = get_data(path)
-    continuous_measures = CONTINUOUS_MEASURES(data)
-    distribution = GAMMA(continuous_measures=continuous_measures)
+    continuous_measures = ContinuousMeasures(data)
+    distribution = Gamma(continuous_measures=continuous_measures)
 
     print(f"{distribution.name} distribution")
     print(f"Parameters: {distribution.parameters}")
