+++ phitter_web/discrete/phitter_web_discrete.py
@@ -1186,7 +1186,6 @@ class PhitterDiscrete:
         plot_height: int,
         plot_width: int,
         plot_bar_color: str,
-        plot_bargap: float,
     ):
         plt.style.use("ggplot")
         plt.style.use("ggplot")
