+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "phitter"
-version = "1.0.0"
+version = "1.0.1"
 description = "Find the best probability distribution for your dataset and simulate processes and queues"
 authors = [{name = "Sebastián José Herrera Monterrosa",  email = "phitter.email@gmail.com"}, {name = "Carlos Andrés Másmela Pinilla",  email = "phitter.email@gmail.com"}]
 readme = "README.md"

[…]
