### 2025-03-25 d54b9f2
add items to distributions descriptions to improve reading
