+++ phitter/discrete/discrete_statistical_tests/discrete_test_kolmogorov_smirnov.py
@@ -1,5 +1,3 @@
-import sys
-
 import numpy
 import scipy.stats
 
