+++ distributions_documentation/continuous/document_continuous_distributions/phitter_continuous_distributions.tex
@@ -80,7 +80,7 @@
 \end{itemize}
 \subsection{Spreadsheet documents}
 \begin{itemize}
-    \item \href{https://github.com/phitterio/phitter-files/blob/main/continuous/alpha.xlsx}{\color{teal}{Excel file from GitHub repository}}
+    \item \href{https://github.com/phitter-core/phitter-files/blob/main/continuous/alpha.xlsx}{\color{teal}{Excel file from GitHub repository}}
     \item \href{https://docs.google.com/spreadsheets/d/1yRovxx1YbqgEul65DjjXetysc_4qgX2a_2NQQA1AxCA}{\color{teal}{Google spreadsheet document}}
 \end{itemize}
 
@@ -123,7 +123,7 @@
 \end{itemize}
 \subsection{Spreadsheet documents}
 \begin{itemize}
-    \item \href{https://github.com/phitterio/phitter-files/blob/main/continuous/arcsine.xlsx}{\color{teal}{Excel file from GitHub repository}}
+    \item \href{https://github.com/phitter-core/phitter-files/blob/main/continuous/arcsine.xlsx}{\color{teal}{Excel file from GitHub repository}}
     \item \href{https://docs.google.com/spreadsheets/d/1q8SKX4gmSbpGzimRvjopzaZ4KrEV5NY1EPmf1G1T7NQ}{\color{teal}{Google spreadsheet document}}
 \end{itemize}
 
@@ -175,7 +175,7 @@
 \end{itemize}
 \subsection{Spreadsheet documents}
 \begin{itemize}
-    \item \href{https://github.com/phitterio/phitter-files/blob/main/continuous/argus.xlsx}{\color{teal}{Excel file from GitHub repository}}
+    \item \href{https://github.com/phitter-core/phitter-files/blob/main/continuous/argus.xlsx}{\color{teal}{Excel file from GitHub repository}}
     \item \href{https://docs.google.com/spreadsheets/d/1u2x7IFUSB7rEyhs7s6-C2btT1Bk5aCr4WiUYEML-8xs}{\color{teal}{Google spreadsheet document}}
 \end{itemize}
 
@@ -221,7 +221,7 @@
 \end{itemize}
 \subsection{Spreadsheet documents}
 \begin{itemize}
-    \item \href{https://github.com/phitterio/phitter-files/blob/main/continuous/beta.xlsx}{\color{teal}{Excel file from GitHub repository}}
+    \item \href{https://github.com/phitter-core/phitter-files/blob/main/continuous/beta.xlsx}{\color{teal}{Excel file from GitHub repository}}
     \item \href{https://docs.google.com/spreadsheets/d/1P7NDy-9toV3dv64gabnr8l2NjB1xt_Ani5IVMTx3gyU}{\color{teal}{Google spreadsheet document}}
 \end{itemize}
 
@@ -266,7 +266,7 @@
 \end{itemize}
 \subsection{Spreadsheet documents}
 \begin{itemize}
-    \item \href{https://github.com/phitterio/phitter-files/blob/main/continuous/beta_prime.xlsx}{\color{teal}{Excel file from GitHub repository}}
+    \item \href{https://github.com/phitter-core/phitter-files/blob/main/continuous/beta_prime.xlsx}{\color{teal}{Excel file from GitHub repository}}
     \item \href{https://docs.google.com/spreadsheets/d/1-8cKeS9D6YixQE_uLig7UarXcoQoE-341yHDj8sfXA8}{\color{teal}{Google spreadsheet document}}
 \end{itemize}
 
@@ -315,7 +315,7 @@
 \end{itemize}
 \subsection{Spreadsheet documents}
 \begin{itemize}
[…]
