+++ documentation/FIT.md
@@ -93,7 +93,7 @@ phi.fit()
 phi.plot_histogram()
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitterio/phitter-kernel/blob/main/multimedia/plot_histogram.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_histogram.png?raw=true" width="500" />
 
 ### Histogram PDF Distributions Plot
 
@@ -107,7 +107,7 @@ phi.fit()
 phi.plot_histogram_distributions()
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitterio/phitter-kernel/blob/main/multimedia/plot_histogram_distributions.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_histogram_distributions.png?raw=true" width="500" />
 
 ### Histogram PDF Distribution Plot
 
@@ -121,7 +121,7 @@ phi.fit()
 phi.plot_distribution("beta")
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitterio/phitter-kernel/blob/main/multimedia/plot_one_distribution.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_one_distribution.png?raw=true" width="500" />
 
 ### ECDF Plot
 
@@ -135,7 +135,7 @@ phi.fit()
 phi.plot_ecdf()
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitterio/phitter-kernel/blob/main/multimedia/plot_ecdf.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_ecdf.png?raw=true" width="500" />
 
 ### ECDF Distribution Plot
 
@@ -149,7 +149,7 @@ phi.fit()
 phi.plot_ecdf_distribution("beta")
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitterio/phitter-kernel/blob/main/multimedia/plot_ecdf_distribution.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_ecdf_distribution.png?raw=true" width="500" />
 
 ### QQ Plot
 
@@ -163,7 +163,7 @@ phi.fit()
 phi.qq_plot("beta")
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitterio/phitter-kernel/blob/main/multimedia/plot_qq.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_qq.png?raw=true" width="500" />
 
 ### QQ - Regression Plot
 
@@ -177,7 +177,7 @@ phi.fit()
 phi.qq_plot_regression("beta")
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitterio/phitter-kernel/blob/main/multimedia/plot_qq_regression.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_qq_regression.png?raw=true" width="500" />
 
 ### Working with distributions: Methods and properties
 
@@ -204,102 +204,102 @@ distribution.mode # -> 733.3333333333333
 
 ## Continuous Distributions
 
[…]
