### 2025-04-07 49e549f
update JOSS paper, update pyproject file
