+++ paper/paper.bib
@@ -49,3 +49,42 @@
   year      = {1961},
   publisher = {JSTOR}
 }
+
+@book{vose2008risk,
+  title     = {Risk analysis: a quantitative guide},
+  author    = {Vose, David},
+  year      = {2008},
+  publisher = {John Wiley \& Sons}
+}
+
+@article{2020SciPy-NMeth,
+  author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
+             Haberland, Matt and Reddy, Tyler and Cournapeau, David and
+             Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and
+             Bright, Jonathan and {van der Walt}, St{\'e}fan J. and
+             Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and
+             Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and
+             Kern, Robert and Larson, Eric and Carey, C J and
+             Polat, {\.I}lhan and Feng, Yu and Moore, Eric W. and
+             {VanderPlas}, Jake and Laxalde, Denis and Perktold, Josef and
+             Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and
+             Harris, Charles R. and Archibald, Anne M. and
+             Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
+             {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
+  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
+             Computing in Python}},
+  journal = {Nature Methods},
+  year    = {2020},
+  volume  = {17},
+  pages   = {261--272},
+  adsurl  = {https://rdcu.be/b08Wh},
+  doi     = {10.1038/s41592-019-0686-2}
+}
+
+@misc{wiki:List_of_probability_distributions,
+  author       = {Wikipedia},
+  title        = {{List of probability distributions} --- {W}ikipedia{,} The Free Encyclopedia},
+  year         = {2025},
+  howpublished = {\url{http://en.wikipedia.org/w/index.php?title=List\%20of\%20probability\%20distributions&oldid=1282462828}},
+  note         = {[Online; accessed 06-April-2025]}
+}
