+++ paper/paper.md
@@ -3,17 +3,20 @@ title: "Phitter: A Python Library for Probability Distribution Fitting and Analy
 tags:
     - Python
     - Statistics
-    - probability distributions
-    - data analysis
-    - machine learning
-    - simulation
-    - monte carlo
+    - Probability Distributions
+    - Data Analysis
+    - Machine Learning
+    - Simulation
+    - Monte Carlo
 authors:
     - name: Sebastián José Herrera Monterrosa
       orcid: 0009-0002-2766-642X
       affiliation: 1
+    - name: Carlos Andrés Masmela Pinilla
+      orcid: 0009-0000-0390-1558
+      affiliation: 1
 affiliations:
-    - name: Pontificia Universidad Javeriana
+    - name: "Pontificia Universidad Javeriana"
       index: 1
 date: 26 March 2024
 bibliography: paper.bib
@@ -21,53 +24,52 @@ bibliography: paper.bib
 
 # Summary
 
-Phitter is a Python library designed to analyze datasets and determine the best analytical probability distributions that represent them. It provides a comprehensive suite of tools for fitting and analyzing over 80 probability distributions, both continuous and discrete. Phitter implements three goodness-of-fit tests and offers interactive visualizations to aid in the analysis process. For each selected probability distribution, Phitter provides a standard modeling guide along with detailed spreadsheets that outline the methodology for using the chosen distribution in various fields such as data science, operations research, and artificial intelligence.
+Phitter is an open-source Python library that streamlines the process of fitting and analyzing probability distributions across various domains—including statistics, data science, operations research, and machine learning—by offering a comprehensive catalog of over 80 continuous and discrete distributions, multiple goodness-of-fit measures (Chi-Square, Kolmogorov-Smirnov, and Anderson-Darling), interactive visualizations for exploratory data analysis and model validation, and detailed modeling guides and spreadsheets to assist users in applying the chosen distribution for simulation or predictive tasks, thus reducing the complexity of distribution fitting and helping researchers and practitioners identify distributions that most closely model their data.
 
 # Statement of Need
 
-In the fields of data science, statistics, and machine learning, understanding the underlying probability distributions of datasets is crucial for accurate modeling and prediction. However, identifying the most appropriate distribution for a given dataset can be a complex and time-consuming task. Phitter addresses this need by providing a user-friendly, efficient, and comprehensive tool for probability distribution fitting and analysis.
+Fitting probability distributions to empirical data is a fundamental task in various scientific and engineering disciplines, crucial for applications such as stochastic modeling, risk assessment, and event simulation [@vose2008risk]. While methods exist for fitting specific distributions where parameters are easily derived from sample statistics (e.g., the normal distribution), the process becomes significantly more complex when aiming to identify the best-fitting distribution among a large set of candidates or when dealing with distributions requiring non-trivial parameter estimation techniques.
 
-Phitter stands out from existing tools by offering:
+Although commercial software packages exist that automate distribution fitting, their commercial nature and cost limit accessibility for many researchers and practitioners. Within the open-source scientific Python ecosystem, the `scipy.stats` module [@2020SciPy-NMeth] offers a powerful foundation, providing implementations for numerous probability distributions and parameter estimation capabilities.
 
-1. A wide range of over 80 probability distributions, including both continuous and discrete options.
-2. Implementation of multiple goodness-of-fit tests (Chi-Square, Kolmogorov-Smirnov, and Anderson-Darling).
-3. Interactive visualizations for better understanding and interpretation of results.
-4. Accelerated fitting capabilities for large datasets (over 100K samples).
[…]
