### 2025-04-07 #7625 @sebastianherreramonterrosa
Hey @kanishkan91, sorry for the delayed response! I've been working on the library documentation, which covers most of the remaining items. You can check it out here: [Phitter docs](https://docs-phitter-kernel.netlify.app/).

Additionally, I've updated the paper to address the previously raised concerns:

1. I have clarified the features and scope of the library.
2. I have included @candemas97 as a co-author.
3. I have added comments comparing our library to similar ones.
4. I have made a few other minor improvements.

Let me know if you need anything else!
