### 2025-04-08 #7625 @sebastianherreramonterrosa
@EwoutH, here the answers for your comments.

> 1. Contribution and authorship: It seems Carlos Másmela has made significant contributions to the repo. Should he be listed as a co-author?

I have already included @candemas97 as a co-author.

> 2. Functionality
> a. The paper claims "over 80 probability distributions". I think I found them here: https://github.com/phitterio/phitter-kernel/blob/main/documentation/FIT.md#continuous-distributions, but it might be nice to hyperlink to the documentation in the paper so users can find this list.

You can find a comprehensive list of these distributions in the following sections of the documentation:
- https://docs-phitter-kernel.netlify.app/documentation/distributions/continuous-distributions.html
- https://docs-phitter-kernel.netlify.app/documentation/distributions/discrete-distributions.html
- Additionally, there is a dedicated page for each distribution, such as [Exponential Distribution](https://docs-phitter-kernel.netlify.app/documentation/distributions/continuous/exponential.html).

>    b. "Accelerated fitting capabilities for large datasets (over 100K samples)" --> Where can I find an example of this?

You can find an explnaition and example in the documentation here:
- https://docs-phitter-kernel.netlify.app/documentation/fit/accelerate/speed-up-large-datasets.html

> 3. Documentation
>    a. Installation instructions --> Maybe link to docs here, especially "Is there a clearly-stated list of dependencies?" seems to be a JOSS requirement ([@kanishkan91](https://github.com/kanishkan91) does that need to be in the main paper, or can that be linked to in docs?)

In the section of installation are included the dependencies.
- https://docs-phitter-kernel.netlify.app/documentation/intro/installation.html

>    b. Example usage is not present. A few simple examples (like in the Readme or docs) might be useful.

Here are some examples:

- https://docs-phitter-kernel.netlify.app/documentation/intro/examples.html
- https://docs-phitter-kernel.netlify.app/documentation/fit/quick-fit.html
- https://docs-phitter-kernel.netlify.app/documentation/fit/continuous/continuous-fit-implementation.html
- https://docs-phitter-kernel.netlify.app/documentation/fit/discrete/discrete-fit-implementation.html

>    c. Functionality documentation --> Idem dito ([@kanishkan91](https://github.com/kanishkan91) I thought JOSS was [against](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) API documentation? What's the balance here?
>    _Given this format, a “full length” paper is not permitted, and software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation._

The documentation is ready and the paper has been updated.

>    d. Community guidelines --> [@kanishkan91](https://github.com/kanishkan91) is [this](https://github.com/phitterio/phitter-kernel?tab=readme-ov-file#contribution) enough or does JOSS require more?

The readme has been updated for this item.

> 4. Software paper
>    a. A statement of need / State of the field --> Might be nice to name the largest competitors of Phitter
>    b. References --> There are currently none, it might be nice to refer to statistical foundations, related Python statistical libraries, other comparative software and applications.

The revised paper now includes a statement of need, mentioning the major competitors of Phitter to better contextualize its relevance.. The refereces have been updated.
