### 2025-04-08 #7625 @sebastianherreramonterrosa
@mdhaber, here the answers for the topics remain to be answered.

1. The references for the definitions of the distributions have been included in the paper.
2. The comparisons with similar libraries have been included in the new paper version.
3. Thinking about the speed of the fit, evaluating the Anderson-Darling distribution using the Marsaglia & Marsaglia approach seems like a good decision to me.
