+++ phitter/continuous/continuous_statistical_tests/continuous_test_chi_square.py
@@ -51,7 +51,12 @@ def evaluate_continuous_test_chi_square(distribution, continuous_measures):
     rejected = statistic_chi2 >= critical_value
 
     ## Construction of answer
-    result_test_chi2 = {"test_statistic": statistic_chi2, "critical_value": critical_value, "p-value": p_value, "rejected": rejected}
+    result_test_chi2 = {
+        "test_statistic": statistic_chi2,
+        "critical_value": critical_value,
+        "p-value": p_value,
+        "rejected": rejected,
+    }
     return result_test_chi2
 
 
