+++ phitter/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.0.2"
+__version__ = "1.0.3"
 
 from .main import Phitter
 from phitter import continuous
