### 2025-04-17 #7625 @mdhaber
I think the changes are good. A few typos:

- "Implements efficient parameter estimation techniques, prioritizing **th** time estimation."
- `Phitter implements probability distributions as standardized as possible according to the following sources: [@walck1996hand], [@mclaughlin2001compendium], [@wiki:List_of_probability_distributions]` - I think that should be semicolon-separated within a single set of brackets. [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)
- "Phitter *evaluetes de Anderson Darling* distribution" - "evaluates the Anderson-Darling test"?
- "proveides Excel and Google sheet **implemetation**"

Other thoughts about the revised paper:

- The summary is all one sentence. It appears to be grammatically correct, but it might be easier to read if broken up.
- At the end of "providing implementations for numerous probability distributions and parameter estimation capabilities.", I'd recommend adding a phrase ", but..." that specifies what SciPy is missing (e.g. each distribution must identified and fitted separately, fitting is too slow, etc.). Right now, "Therefore, there is a need..." felt a bit like a non sequitur to me.
- Thanks for comparing to existing software. The Statement of Need might be a little stronger if those were compared up there, right after SciPy. I'd also recommend adding them to the references - even their GH repo, if they haven't published anything.

The remaining elements of my checklist are:
- [ ] Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)? *There is extensive documentation, but it does not appear to have "API method documentation", at least not as distinct from examples/tutorials as discussed above. This is fine with me if it's OK with the@EwoutH and editor @kanishkan91.*
- [ ] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)? *See above.*
