### 2025-04-17 #7625 @mdhaber
@EwoutH Let's make sure we agree on what "API documentation" is so we can provide the authors with clear direction. To distinguish between that the current documentation has (examples) and what we're asking for, I mentioned "Please see the distinction between the [User Guide](https://scikit-learn.org/stable/user_guide.html) and [API Reference](https://scikit-learn.org/stable/api/index.html) of scikit-learn" above. Is that what you have in mind, too?
