### 2025-04-17 #7625 @kanishkan91
@mdhaber @EwoutH - Thanks for the detailed review. I think API dcoumentation as what JOSS sees it is a general documentation of the methods, classes. This is usally the same that is generated automatically when "docs" are generated for a python package for example.

@sebastianherreramonterrosa - I would suggest that you just generate the docs for the phitter python code using sphinx or any other package like that and just host that on your repo. By doing this we would have all the classes and methods documented somewhere. As a simple example, here's what the docs look like for a package -https://jgcri.github.io/demeter/build/html/demeter.html

@mdhaber @EwoutH- What do you think?
