### 2025-04-17 #7625 @mdhaber
Yeah, I think we're on the same page. It does look like at least some of the functions have a summary and descriptions of arguments and return values. It's a matter of completing that and getting it to render.
