### 2025-04-17 #7625 @EwoutH
Yeah we’re 100% on the same page. Just Classes, methods and their docstring rendered (completely and consistently).
