+++ api_documentation/markdown/modules.md
@@ -1,0 +1,60 @@
+* phitter
+* [View page source](_sources/modules.rst.txt)
+
+---
+
+# phitter[](#phitter "Link to this heading")
+
+* [phitter package](phitter.html)
++ [Subpackages](phitter.html#subpackages)
+## - [phitter.continuous package](phitter.continuous.html)
+## * [Subpackages](phitter.continuous.html#subpackages)
+## * [Submodules](phitter.continuous.html#submodules)
+## * [phitter.continuous.phitter\_continuous module](phitter.continuous.html#module-phitter.continuous.phitter_continuous)
+## * [Module contents](phitter.continuous.html#module-phitter.continuous)
+## - [phitter.discrete package](phitter.discrete.html)
+## * [Subpackages](phitter.discrete.html#subpackages)
+## * [Submodules](phitter.discrete.html#submodules)
+## * [phitter.discrete.phitter\_discrete module](phitter.discrete.html#module-phitter.discrete.phitter_discrete)
+## * [Module contents](phitter.discrete.html#module-phitter.discrete)
+## - [phitter.simulation package](phitter.simulation.html)
+## * [Subpackages](phitter.simulation.html#subpackages)
+## * [Module contents](phitter.simulation.html#module-phitter.simulation)
++ [Submodules](phitter.html#submodules)
++ [phitter.main module](phitter.html#module-phitter.main)
+## - [`Phitter`](phitter.html#phitter.main.Phitter)
+## * [`Phitter.best_distribution`](phitter.html#phitter.main.Phitter.best_distribution)
+## * [`Phitter.df_not_rejected_distributions`](phitter.html#phitter.main.Phitter.df_not_rejected_distributions)
+## * [`Phitter.df_sorted_distributions_sse`](phitter.html#phitter.main.Phitter.df_sorted_distributions_sse)
+## * [`Phitter.dict_to_dataframe()`](phitter.html#phitter.main.Phitter.dict_to_dataframe)
+## * [`Phitter.fit()`](phitter.html#phitter.main.Phitter.fit)
+## * [`Phitter.get_n_test_null()`](phitter.html#phitter.main.Phitter.get_n_test_null)
+## * [`Phitter.get_n_test_passed()`](phitter.html#phitter.main.Phitter.get_n_test_passed)
+## * [`Phitter.get_parameters()`](phitter.html#phitter.main.Phitter.get_parameters)
+## * [`Phitter.get_sse()`](phitter.html#phitter.main.Phitter.get_sse)
+## * [`Phitter.get_test_anderson_darling()`](phitter.html#phitter.main.Phitter.get_test_anderson_darling)
+## * [`Phitter.get_test_chi_square()`](phitter.html#phitter.main.Phitter.get_test_chi_square)
+## * [`Phitter.get_test_kolmogorov_smirnov()`](phitter.html#phitter.main.Phitter.get_test_kolmogorov_smirnov)
+## * [`Phitter.not_rejected_distributions`](phitter.html#phitter.main.Phitter.not_rejected_distributions)
+## * [`Phitter.plot_distribution()`](phitter.html#phitter.main.Phitter.plot_distribution)
+## * [`Phitter.plot_ecdf()`](phitter.html#phitter.main.Phitter.plot_ecdf)
+## * [`Phitter.plot_ecdf_distribution()`](phitter.html#phitter.main.Phitter.plot_ecdf_distribution)
+## * [`Phitter.plot_histogram()`](phitter.html#phitter.main.Phitter.plot_histogram)
+## * [`Phitter.plot_histogram_distributions()`](phitter.html#phitter.main.Phitter.plot_histogram_distributions)
[…]
