+++ api_documentation/markdown/phitter.md
@@ -1,0 +1,267 @@
+* [phitter](modules.html)
+* phitter package
+* [View page source](_sources/phitter.rst.txt)
+
+---
+
+# phitter package[](#phitter-package "Link to this heading")
+
+## Subpackages[](#subpackages "Link to this heading")
+
+* [phitter.continuous package](phitter.continuous.html)
++ [Subpackages](phitter.continuous.html#subpackages)
+## - [phitter.continuous.continuous\_distributions package](phitter.continuous.continuous_distributions.html)
+## * [Submodules](phitter.continuous.continuous_distributions.html#submodules)
+## * [phitter.continuous.continuous\_distributions.alpha module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.alpha)
+## * [phitter.continuous.continuous\_distributions.arcsine module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.arcsine)
+## * [phitter.continuous.continuous\_distributions.argus module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.argus)
+## * [phitter.continuous.continuous\_distributions.beta module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.beta)
+## * [phitter.continuous.continuous\_distributions.beta\_prime module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.beta_prime)
+## * [phitter.continuous.continuous\_distributions.beta\_prime\_4p module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.beta_prime_4p)
+## * [phitter.continuous.continuous\_distributions.bradford module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.bradford)
+## * [phitter.continuous.continuous\_distributions.burr module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.burr)
+## * [phitter.continuous.continuous\_distributions.burr\_4p module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.burr_4p)
+## * [phitter.continuous.continuous\_distributions.cauchy module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.cauchy)
+## * [phitter.continuous.continuous\_distributions.chi\_square module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.chi_square)
+## * [phitter.continuous.continuous\_distributions.chi\_square\_3p module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.chi_square_3p)
+## * [phitter.continuous.continuous\_distributions.dagum module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.dagum)
+## * [phitter.continuous.continuous\_distributions.dagum\_4p module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.dagum_4p)
+## * [phitter.continuous.continuous\_distributions.erlang module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.erlang)
+## * [phitter.continuous.continuous\_distributions.erlang\_3p module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.erlang_3p)
+## * [phitter.continuous.continuous\_distributions.error\_function module](phitter.continuous.continuous_distributions.html#module-phitter.continuous.continuous_distributions.error_function)
[…]
