+++ api_documentation/markdown/phitter.discrete.discrete_measures.md
@@ -1,0 +1,35 @@
+* [phitter](modules.html)
+* [phitter package](phitter.html)
+* [phitter.discrete package](phitter.discrete.html)
+* phitter.discrete.discrete\_measures package
+* [View page source](_sources/phitter.discrete.discrete_measures.rst.txt)
+
+---
+
+# phitter.discrete.discrete\_measures package[](#phitter-discrete-discrete-measures-package "Link to this heading")
+
+## Submodules[](#submodules "Link to this heading")
+
+## phitter.discrete.discrete\_measures.discrete\_measures module[](#module-phitter.discrete.discrete_measures.discrete_measures "Link to this heading")
+
+*class* phitter.discrete.discrete\_measures.discrete\_measures.DiscreteMeasures(*data*, *confidence\_level=0.95*, *subsample\_size=None*, *subsample\_estimation\_size=None*)[](#phitter.discrete.discrete_measures.discrete_measures.DiscreteMeasures "Link to this definition")
+:   Bases: `object`
+
+## critical\_value\_chi2(*freedom\_degrees*)[](#phitter.discrete.discrete_measures.discrete_measures.DiscreteMeasures.critical_value_chi2 "Link to this definition")
+
+## get\_dict()[](#phitter.discrete.discrete_measures.discrete_measures.DiscreteMeasures.get_dict "Link to this definition")
+Return type
+`str`
+
+## Module contents[](#module-phitter.discrete.discrete_measures "Link to this heading")
+
+[Previous](phitter.discrete.discrete_distributions.html "phitter.discrete.discrete_distributions package")
+[Next](phitter.discrete.discrete_statistical_tests.html "phitter.discrete.discrete_statistical_tests package")
+
+---
+
+© Copyright .
+
+Built with [Sphinx](https://www.sphinx-doc.org/) using a
+[theme](https://github.com/readthedocs/sphinx_rtd_theme)
+provided by [Read the Docs](https://readthedocs.org).

[…]
