+++ api_documentation/markdown/index.md
@@ -3,7 +3,7 @@
 
 ---
 
-# phitter documentation[](#phitter-documentation "Link to this heading")
+# phitter documentation
 
 Add your content using `reStructuredText` syntax. See the
 [reStructuredText](https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html)
