+++ api_documentation/markdown/phitter.discrete.discrete_distributions.md
@@ -6,32 +6,32 @@
 
 ---
 
-# phitter.discrete.discrete\_distributions package[](#phitter-discrete-discrete-distributions-package "Link to this heading")
+# phitter.discrete.discrete\_distributions package
 
-## Submodules[](#submodules "Link to this heading")
+## Submodules
 
-## phitter.discrete.discrete\_distributions.bernoulli module[](#module-phitter.discrete.discrete_distributions.bernoulli "Link to this heading")
+## phitter.discrete.discrete\_distributions.bernoulli module
 
-*class* phitter.discrete.discrete\_distributions.bernoulli.Bernoulli(*parameters=None*, *discrete\_measures=None*, *init\_parameters\_examples=False*)[](#phitter.discrete.discrete_distributions.bernoulli.Bernoulli "Link to this definition")
+*class* phitter.discrete.discrete\_distributions.bernoulli.Bernoulli(*parameters=None*, *discrete\_measures=None*, *init\_parameters\_examples=False*)
 :   Bases: `object`
 
 ## Bernoulli distribution
 ## - Parameters Bernoulli Distribution: {“p”: \*}
 ## - <https://phitter.io/distributions/discrete/bernoulli>
 
-## cdf(*x*)[](#phitter.discrete.discrete_distributions.bernoulli.Bernoulli.cdf "Link to this definition")
+## cdf(*x*)
 Cumulative distribution function
 
 #### Return type
 `float` | `ndarray`
 
-## central\_moments(*k*)[](#phitter.discrete.discrete_distributions.bernoulli.Bernoulli.central_moments "Link to this definition")
+## central\_moments(*k*)
 Parametric central moments. µ’[k] = E[(X - E[X])ᵏ] = ∫(x-µ[k])ᵏ∙f(x) dx
 
 #### Return type
 `float` | `None`
 
-## get\_parameters(*discrete\_measures*)[](#phitter.discrete.discrete_distributions.bernoulli.Bernoulli.get_parameters "Link to this definition")
+## get\_parameters(*discrete\_measures*)
 Calculate proper parameters of the distribution from sample discrete\_measures.
 - The parameters are calculated by formula.
 
@@ -44,86 +44,86 @@ Calculate proper parameters of the distribution from sample discrete\_measures.
 #### Return type
 {“p”: \*}
 
-## *property* kurtosis*: float*[](#phitter.discrete.discrete_distributions.bernoulli.Bernoulli.kurtosis "Link to this definition")
+## *property* kurtosis*: float*
 Parametric kurtosis
 
-## *property* mean*: float*[](#phitter.discrete.discrete_distributions.bernoulli.Bernoulli.mean "Link to this definition")
+## *property* mean*: float*
 Parametric mean
 
-## *property* median*: float*[](#phitter.discrete.discrete_distributions.bernoulli.Bernoulli.median "Link to this definition")
+## *property* median*: float*
 Parametric median
 
-## *property* mode*: float*[](#phitter.discrete.discrete_distributions.bernoulli.Bernoulli.mode "Link to this definition")
+## *property* mode*: float*
 Parametric mode
 
-## *property* name[](#phitter.discrete.discrete_distributions.bernoulli.Bernoulli.name "Link to this definition")
+## *property* name
 
-## non\_central\_moments(*k*)[](#phitter.discrete.discrete_distributions.bernoulli.Bernoulli.non_central_moments "Link to this definition")
+## non\_central\_moments(*k*)
 Parametric no central moments. µ[k] = E[Xᵏ] = ∫xᵏ∙f(x) dx
 
 #### Return type
 `float` | `None`
 
-## *property* num\_parameters*: int*[](#phitter.discrete.discrete_distributions.bernoulli.Bernoulli.num_parameters "Link to this definition")
+## *property* num\_parameters*: int*
[…]
