+++ api_documentation/markdown/phitter.continuous.md
@@ -5,9 +5,9 @@
 
 ---
 
-# phitter.continuous package[](#phitter-continuous-package "Link to this heading")
+# phitter.continuous package
 
-## Subpackages[](#subpackages "Link to this heading")
+## Subpackages
 
 * [phitter.continuous.continuous\_distributions package](phitter.continuous.continuous_distributions.html)
 + [Submodules](phitter.continuous.continuous_distributions.html#submodules)
@@ -1495,52 +1495,52 @@
 ## - [`evaluate_continuous_test_kolmogorov_smirnov()`](phitter.continuous.continuous_statistical_tests.html#phitter.continuous.continuous_statistical_tests.continuous_test_kolmogorov_smirnov.evaluate_continuous_test_kolmogorov_smirnov)
 + [Module contents](phitter.continuous.continuous_statistical_tests.html#module-phitter.continuous.continuous_statistical_tests)
 
-## Submodules[](#submodules "Link to this heading")
+## Submodules
 
-## phitter.continuous.phitter\_continuous module[](#module-phitter.continuous.phitter_continuous "Link to this heading")
+## phitter.continuous.phitter\_continuous module
 
-*class* phitter.continuous.phitter\_continuous.PhitterContinuous(*data*, *num\_bins=None*, *confidence\_level=0.95*, *minimum\_sse=inf*, *subsample\_size=None*, *subsample\_estimation\_size=None*, *distributions\_to\_fit='all'*, *exclude\_distributions='any'*)[](#phitter.continuous.phitter_continuous.PhitterContinuous "Link to this definition")
+*class* phitter.continuous.phitter\_continuous.PhitterContinuous(*data*, *num\_bins=None*, *confidence\_level=0.95*, *minimum\_sse=inf*, *subsample\_size=None*, *subsample\_estimation\_size=None*, *distributions\_to\_fit='all'*, *exclude\_distributions='any'*)
 :   Bases: `object`
 
-## fit(*n\_workers=1*)[](#phitter.continuous.phitter_continuous.PhitterContinuous.fit "Link to this definition")
+## fit(*n\_workers=1*)
 
-## parse\_rgba\_color(*rgba\_string*)[](#phitter.continuous.phitter_continuous.PhitterContinuous.parse_rgba_color "Link to this definition")
+## parse\_rgba\_color(*rgba\_string*)
 
-## plot\_distribution\_pdf\_matplotlib(*id\_distribution*, *plot\_title*, *plot\_xaxis\_title*, *plot\_yaxis\_title*, *plot\_legend\_title*, *plot\_height*, *plot\_width*, *plot\_bar\_color*, *plot\_bargap*, *plot\_line\_color*, *plot\_line\_width*)[](#phitter.continuous.phitter_continuous.PhitterContinuous.plot_distribution_pdf_matplotlib "Link to this definition")
+## plot\_distribution\_pdf\_matplotlib(*id\_distribution*, *plot\_title*, *plot\_xaxis\_title*, *plot\_yaxis\_title*, *plot\_legend\_title*, *plot\_height*, *plot\_width*, *plot\_bar\_color*, *plot\_bargap*, *plot\_line\_color*, *plot\_line\_width*)
 
[…]
