+++ api_documentation/markdown/phitter.simulation.own_distribution.md
@@ -6,16 +6,16 @@
 
 ---
 
-# phitter.simulation.own\_distribution package[](#phitter-simulation-own-distribution-package "Link to this heading")
+# phitter.simulation.own\_distribution package
 
-## Submodules[](#submodules "Link to this heading")
+## Submodules
 
-## phitter.simulation.own\_distribution.own\_distribution module[](#module-phitter.simulation.own_distribution.own_distribution "Link to this heading")
+## phitter.simulation.own\_distribution.own\_distribution module
 
-*class* phitter.simulation.own\_distribution.own\_distribution.OwnDistributions(*parameters*)[](#phitter.simulation.own_distribution.own_distribution.OwnDistributions "Link to this definition")
+*class* phitter.simulation.own\_distribution.own\_distribution.OwnDistributions(*parameters*)
 :   Bases: `object`
 
-## ppf(*probability*)[](#phitter.simulation.own_distribution.own_distribution.OwnDistributions.ppf "Link to this definition")
+## ppf(*probability*)
 Assign a label according to a probability given by the created distribution
 
 #### Parameters
@@ -27,7 +27,7 @@ Returns label according to probability
 #### Return type
 int
 
-## Module contents[](#module-phitter.simulation.own_distribution "Link to this heading")
+## Module contents
 
 [Previous](phitter.simulation.html "phitter.simulation package")
 [Next](phitter.simulation.process_simulation.html "phitter.simulation.process_simulation package")

[…]
