### 2025-04-25 67275db
joss paper: typos, comparison libraries citation
