+++ paper/paper.bib
@@ -88,3 +88,25 @@
   howpublished = {\url{http://en.wikipedia.org/w/index.php?title=List\%20of\%20probability\%20distributions&oldid=1282462828}},
   note         = {[Online; accessed 06-April-2025]}
 }
+
+@software{Taskesen_distfit_is_a_2020,
+  author  = {Taskesen, Erdogan},
+  license = {MIT},
+  month   = jan,
+  title   = {{distfit is a python library for probability density fitting.}},
+  url     = {https://erdogant.github.io/distfit},
+  version = {1.4.0},
+  year    = {2020}
+}
+
+@book{Thomas2024cokelaer,
+  doi       = {10.5281/ZENODO.12514960},
+  publisher = {Zenodo},
+  title     = {cokelaer/fitter: v1.7.1},
+  url       = {https://zenodo.org/doi/10.5281/zenodo.12514960},
+  author    = {Thomas Cokelaer and Arseny Kravchenko and lahdjirayhan and msat59 and Vitor Ferrari and Ashutosh Varma and Brian L and Caio Eadi Stringari and Christian Brueffer and Eike Broda and Elmar Pruesse and Karthikeyan Singaravelan and Stefano Alberto Russo and Zhengyi Li and mark padgham and negodfre},
+  date      = {2024-06-24},
+  year      = {2024},
+  month     = {6},
+  day       = {24}
+}
