+++ paper/paper.md
@@ -24,31 +24,31 @@ bibliography: paper.bib
 
 # Summary
 
-Phitter is an open-source Python library that streamlines the process of fitting and analyzing probability distributions across various domains—including statistics, data science, operations research, and machine learning—by offering a comprehensive catalog of over 80 continuous and discrete distributions, multiple goodness-of-fit measures (Chi-Square, Kolmogorov-Smirnov, and Anderson-Darling), interactive visualizations for exploratory data analysis and model validation, and detailed modeling guides and spreadsheets to assist users in applying the chosen distribution for simulation or predictive tasks, thus reducing the complexity of distribution fitting and helping researchers and practitioners identify distributions that most closely model their data.
+Phitter is an open-source Python library designed to streamline the process of fitting and analyzing probability distributions for applications in statistics, data science, operations research, and machine learning. It provides a comprehensive catalog of over 80 continuous and discrete distributions, multiple goodness-of-fit measures (Chi-Square, Kolmogorov-Smirnov, and Anderson-Darling), interactive visualizations for exploratory data analysis and model validation, and detailed modeling guides with spreadsheet implementations. By reducing the complexity of distribution fitting, Phitter helps researchers and practitioners identify distributions that best model their data.
 
 # Statement of Need
 
-Fitting probability distributions to empirical data is a fundamental task in various scientific and engineering disciplines, crucial for applications such as stochastic modeling, risk assessment, and event simulation [@vose2008risk]. While methods exist for fitting specific distributions where parameters are easily derived from sample statistics (e.g., the normal distribution), the process becomes significantly more complex when aiming to identify the best-fitting distribution among a large set of candidates or when dealing with distributions requiring non-trivial parameter estimation techniques.
+Fitting probability distributions to empirical data is a cornerstone of numerous scientific and engineering disciplines, underpinning applications such as stochastic modeling, risk assessment, and event simulation [@vose2008risk]. While fitting certain distributions, like the normal distribution, is straightforward due to easily derived parameters from sample statistics, the task becomes significantly more complex when selecting the best-fitting distribution from a large set of candidates or when estimating parameters for distributions requiring advanced techniques.
 
-Although commercial software packages exist that automate distribution fitting, their commercial nature and cost limit accessibility for many researchers and practitioners. Within the open-source scientific Python ecosystem, the `scipy.stats` module [@2020SciPy-NMeth] offers a powerful foundation, providing implementations for numerous probability distributions and parameter estimation capabilities.
+Commercial software packages that automate distribution fitting exist, but their cost and proprietary nature restrict access for many researchers and practitioners. Within the open-source scientific Python ecosystem, the `scipy.stats` module [@2020SciPy-NMeth] provides a robust foundation, offering implementations of numerous probability distributions and parameter estimation capabilities. However, it requires users to individually select and fit each distribution, lacking a unified interface to systematically evaluate all distributions against a dataset. Each distribution is defined by shape, location, and scale parameters, and parameter estimation is typically performed using maximum likelihood estimation (MLE). While MLE is highly accurate, it is also computationally intensive, especially when applied across many candidate distributions.
 
-Therefore, there is a need for an accessible, open-source Python library specifically designed for automated probability distribution fitting that:
+To address these challenges, there is a clear need for an accessible, open-source Python library tailored for automated probability distribution fitting that:
 
-1.  Provides implementations for a wide range of common distributions.
-2.  Implements efficient parameter estimation techniques, prioritizing th time estimation.
[…]
