+++ phitter/continuous/continuous_distributions/pert.py
@@ -226,7 +226,7 @@ if __name__ == "__main__":
         return data
 
     ## Distribution class
-    path = "../continuous_distributions_sample/sample_pert.txt"
+    path = "../../../distributions_samples/continuous_distributions_sample/sample_pert.txt"
     data = get_data(path)
     continuous_measures = ContinuousMeasures(data)
     distribution = Pert(continuous_measures=continuous_measures)
