+++ phitter/continuous/continuous_distributions/fatigue_life.py
@@ -216,7 +216,7 @@ if __name__ == "__main__":
         return data
 
     ## Distribution class
-    path = "../continuous_distributions_sample/sample_fatigue_life.txt"
+    path = "../../../distributions_samples/continuous_distributions_sample/sample_fatigue_life.txt"
     data = get_data(path)
     continuous_measures = ContinuousMeasures(data)
     distribution = FatigueLife(continuous_measures=continuous_measures)
