+++ phitter/continuous/continuous_distributions/weibull_3p.py
@@ -233,7 +233,7 @@ if __name__ == "__main__":
         return data
 
     ## Distribution class
-    path = "../continuous_distributions_sample/sample_weibull_3p.txt"
+    path = "../../../distributions_samples/continuous_distributions_sample/sample_weibull_3p.txt"
     data = get_data(path)
     continuous_measures = ContinuousMeasures(data)
     distribution = Weibull3P(continuous_measures=continuous_measures)
