+++ phitter/continuous/continuous_distributions/johnson_sb.py
@@ -246,7 +246,7 @@ if __name__ == "__main__":
         return data
 
     ## Distribution class
-    path = "../continuous_distributions_sample/sample_johnson_sb.txt"
+    path = "../../../distributions_samples/continuous_distributions_sample/sample_johnson_sb.txt"
     data = get_data(path)
     continuous_measures = ContinuousMeasures(data)
     distribution = JohnsonSB(continuous_measures=continuous_measures)
