+++ phitter/continuous/continuous_distributions/pareto_first_kind.py
@@ -254,7 +254,7 @@ if __name__ == "__main__":
         return data
 
     # Distribution class
-    path = "../continuous_distributions_sample/sample_pareto_first_kind.txt"
+    path = "../../../distributions_samples/continuous_distributions_sample/sample_pareto_first_kind.txt"
     data = get_data(path)
     continuous_measures = ContinuousMeasures(data)
     distribution = ParetoFirstKind(continuous_measures=continuous_measures)
