+++ phitter/continuous/continuous_distributions/rice.py
@@ -278,7 +278,7 @@ if __name__ == "__main__":
         return data
 
     ## Distribution class
-    path = "../continuous_distributions_sample/sample_rice.txt"
+    path = "../../../distributions_samples/continuous_distributions_sample/sample_rice.txt"
     data = get_data(path)
     continuous_measures = ContinuousMeasures(data)
     distribution = Rice(continuous_measures=continuous_measures)
