+++ phitter/continuous/continuous_distributions/inverse_gamma.py
@@ -251,7 +251,7 @@ if __name__ == "__main__":
         return data
 
     ## Distribution class
-    path = "../continuous_distributions_sample/sample_inverse_gamma.txt"
+    path = "../../../distributions_samples/continuous_distributions_sample/sample_inverse_gamma.txt"
     data = get_data(path)
     continuous_measures = ContinuousMeasures(data)
     distribution = InverseGamma(continuous_measures=continuous_measures)
