+++ phitter/continuous/continuous_distributions/laplace.py
@@ -183,7 +183,7 @@ if __name__ == "__main__":
         return data
 
     ## Distribution class
-    path = "../continuous_distributions_sample/sample_laplace.txt"
+    path = "../../../distributions_samples/continuous_distributions_sample/sample_laplace.txt"
     data = get_data(path)
     continuous_measures = ContinuousMeasures(data)
     distribution = Laplace(continuous_measures=continuous_measures)
