+++ phitter/continuous/continuous_distributions/non_central_chi_square.py
@@ -230,7 +230,7 @@ if __name__ == "__main__":
         return data
 
     ## Distribution class
-    path = "../continuous_distributions_sample/sample_NON_CENTRAL_CHI_SQUARE.txt"
+    path = "../../../distributions_samples/continuous_distributions_sample/sample_NON_CENTRAL_CHI_SQUARE.txt"
     data = get_data(path)
     continuous_measures = ContinuousMeasures(data)
     distribution = NonCentralChiSquare(continuous_measures=continuous_measures)
