+++ phitter/continuous/continuous_distributions/chi_square.py
@@ -183,7 +183,7 @@ if __name__ == "__main__":
         return data
 
     ## Distribution class
-    path = "../continuous_distributions_sample/sample_chi_square.txt"
+    path = "../../../distributions_samples/continuous_distributions_sample/sample_chi_square.txt"
     data = get_data(path)
     continuous_measures = ContinuousMeasures(data)
     distribution = ChiSquare(continuous_measures=continuous_measures)
