### 2025-04-25 #7625 @sebastianherreramonterrosa
@mdhaber, @EwoutH, @kanishkan91 I have generated the API reference of classes and methods.

Link: https://docs-phitter-kernel-api-reference.netlify.app/modules
