### 2025-04-25 #7625 @sebastianherreramonterrosa
@mdhaber, @EwoutH

1. The typos have been corrected.
2. In the section comparing existing tools, I have included a more detailed description of the mentioned libraries. I have also included the references. Regarding this section, the following question arises: is it necessary to make detailed comparisons, or are the descriptions sufficient? In principle, it is not my intention to make extensive comparisons of the mentioned libraries, unless otherwise specified.
