### 2025-04-28 #7625 @mdhaber
I checked the box for API documentation. This is looking very close.

> Regarding this section, the following question arises: is it necessary to make detailed comparisons, or are the descriptions sufficient?

From my perspective, the primary purpose of mentioning other libraries is to make a case for why *this* library is needed, which is why I suggested that they go in the existing Statement of Need section. @EwoutH, were you asking for them to go in a new section as they do now, or do you agree they should go in the Statement of Need?

_Detailed_ comparisons are not needed, but what is needed IMO is _comparison_. The current version of the text _describes_ the other libraries, but does not highlight differences that motivate why this library still needed. @EwoutH would you agree? Let's again make sure we're on the same page about exactly what needs to happen.

---

The new references are not cited in the text; e.g. `[9, 10, 9]`, `(Taskesen et al., 2020)`, and `(Thomas Cokelaer)` appear to be entered manually, but `[@Taskesen_distfit_is_a_2020]` and `[@Thomas2024cokelaer]` do not appear.
