### 2025-04-28 #7625 @EwoutH
I'm on vacation - so a quick one here:

- Agree that it needs to be in the Statement of Need.
- A "gap" in current libraries should be argued in my opinion. So no full comparison, just why this tool does a (combination of) things no other does (or does something better)

@kanishkan91 I do think there need to be better and more consistent guidelines about this, both for authors and reviewers. Could you put this on the agenda among editors internally?
