### 2025-04-28 #7625 @mdhaber
>  In this context, descriptions would be more than sufficient. Let me know what you think.

My understanding was that the "statement of need" is a required section, and its purpose is to motivate why the package underlying the paper is needed.

If existing packages fully satisfy the needs of users already, then the package underlying the paper is not needed, right?

So to be convincing, I thought that a statement of need should do more than _describe_ other existing packages; it should compare the current package against them to motivate why it is needed.

That said, I only intend to uphold JOSS' standard here, not impose my own, so please correct me if I'm mistaken.
