### 2025-04-28 #7625 @kanishkan91
@mdhaber Thanks ( and I appreciate the discussion on this topic). The statement of need is established from a user perspective. Let's say a user would use a package(to hopefully cite it), they should know through reading the paper why this package was developed and what kind of libraries they should compare it to.

We do not assert "novelty" of any kind. So, if there's a package out there that is faster, that is fine. Authors are encouraged to acknowledge that but not forced to.  Also just as a side note (You are probably aware already!), the idea behind JOSS was that software should be citeable so that scientific software developers can get credit  for their work.

Regardless, @mdhaber @EwoutH - I'l talk to the editors regarding adding more concrete instructions on this. Maybe even refer to this issue directly.
