### 2025-05-13 5adeab6
add comparisons to paper, correction in examples notebooks
