### 2025-05-13 #7625 @sebastianherreramonterrosa
@mdhaber, @EwoutH, @kanishkan91
1. I have corrected the syntax in the reference.
2. I have added comparisons regarding speed and features between Phitter and the libraries being evaluated.
