### 2025-05-14 #7625 @mdhaber
The latest commit corrected the bibliography, so I checked the last box. I cannot verify the new timing data, but if these are not considered "results" that need to be reproducible, I think the paper satisfies the requirements of the checklist.
