### 2025-05-14 #7625 @EwoutH
One last thing: Should you cite `distfit` and `fitter`?
