### 2025-05-16 #7625 @sebastianherreramonterrosa
@kanishkan91. I have uploaded Phitter to Zenodo. Here are the answers for your requests.

1) Version: v1.0.3.JOSS
2) DOI: https://doi.org/10.5281/zenodo.15428773
