+++ README.md
@@ -177,7 +177,7 @@ phi.fit()
 phi.plot_histogram()
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/histogram.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_histogram.png?raw=true" width="500" />
 
 ### Histogram PDF Dsitributions Plot
 
@@ -190,7 +190,7 @@ phi.fit()
 phi.plot_histogram_distributions()
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/histogram_pdf_distributions.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_histogram_distributions.png?raw=true" width="500" />
 
 ### Histogram PDF Dsitribution Plot
 
@@ -203,7 +203,7 @@ phi.fit()
 phi.plot_distribution("beta")
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/histogram_pdf_distribution.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_one_distribution.png?raw=true" width="500" />
 
 ### ECDF Plot
 
@@ -216,7 +216,7 @@ phi.fit()
 phi.plot_ecdf()
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/ecdf.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_ecdf.png?raw=true" width="500" />
 
 ### ECDF Distribution Plot
 
@@ -229,7 +229,7 @@ phi.fit()
 phi.plot_ecdf_distribution("beta")
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/ecdf_distribution.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_ecdf_distribution.png?raw=true" width="500" />
 
 ### QQ Plot
 
@@ -242,7 +242,7 @@ phi.fit()
 phi.qq_plot("beta")
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/qq_plot_distribution.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_qq.png?raw=true" width="500" />
 
 ### QQ - Regression Plot
 
@@ -255,7 +255,7 @@ phi.fit()
 phi.qq_plot_regression("beta")
 ```
 
-<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/qq_plot_distribution_regression.png?raw=true" width="500" />
+<img alt="phitter_histogram" src="https://github.com/phitter-core/phitter-kernel/blob/main/multimedia/plot_qq_regression.png?raw=true" width="500" />
 
 ### Working with distributions: Methods and properties
 
