### 2025-05-19 #7625 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
