### 2025-05-19 8efa45b
joss paper: fix typos
