+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: "Phitter: A library designed to streamline the process of fitting andanalyzing probability distributions"
+title: "Phitter: A library designed to streamline the process of fitting and analyzing probability distributions"
 tags:
     - Python
     - Statistics
@@ -28,7 +28,7 @@ Phitter is an open-source Python library designed to streamline the process of f
 
 # Statement of Need
 
-Fitting probability distributions to empirical data is a cornerstone of numerous scientific and engineering disciplines, underpinning applications such as stochastic modeling, risk assessment, and event simulation [@vose2008risk]. While fitting certain distributions, like the normal distribution, is straightforward due to easily derived parameters from sample statistics, the task becomes significantly more complex when selecting the best-fitting distribution from a large set of candidates or when estimating parameters for distributions requiring advanced techniques.
+Fitting probability distributions to empirical data is a cornerstone of numerous scientific and engineering disciplines, underpinning applications such as stochastic modeling, risk assessment, and event simulation [@vose2008risk]. While fitting certain distributions, like the normal distribution, is straightforward due to easily derived parameters from sample statistics, the task becomes considerably more complex when selecting the most suitable distribution from a large set of candidates or when estimating parameters for distributions requiring advanced techniques.
 
 Commercial software packages that automate distribution fitting exist, but their cost and proprietary nature restrict access for many researchers and practitioners. Within the open-source scientific Python ecosystem, the `scipy.stats` module [@2020SciPy-NMeth] provides a robust foundation, offering implementations of numerous probability distributions and parameter estimation capabilities. However, it requires users to individually select and fit each distribution, lacking a unified interface to systematically evaluate all distributions against a dataset. Each distribution is defined by shape, location, and scale parameters, and parameter estimation is typically performed using maximum likelihood estimation (MLE). While MLE is highly accurate, it is also computationally intensive, especially when applied across many candidate distributions.
 
@@ -44,13 +44,13 @@ To address these challenges, there is a clear need for an accessible, open-sourc
 
 1. Phitter implements probability distributions as standardized as possible according to the following sources: [@walck1996hand; @mclaughlin2001compendium; @wiki:List_of_probability_distributions]
 
-2. **Accelerated Parameter Estimation:** For a subset of implemented distributions, Phitter utilizes direct solving of the distribution's parametric equation system. This approach offers significant computational speed advantages compared to iterative methods like standard Maximum Likelihood Estimation (MLE), particularly beneficial for large datasets. Performance benchmarks detailing estimation times are provided [Estimation Time Parameters for Continuous Distributions](https://docs-phitter-kernel.netlify.app/documentation/benchmarks/continuous/continuous-parameters-estimation.html). Where direct parametric solutions are not implemented or feasible, Phitter seamlessly integrates with and leverages the well-established MLE implementation provided by the SciPy library (`scipy.stats.rv_continuous.fit` and related methods). This ensures broad applicability across a wide range of distributions while prioritizing speed where possible.
[…]
