### 2025-05-30 #7625 @crvernon
:wave: @sebastianherreramonterrosa - I need you to address the following before I move to accept this for publication:

In the archive:

- [ ] Please ensure the title of the archive matches what is in the paper exactly.  For you, this would mean removing "phitter-core/phitter-kernel: ".  There is no need to make a new release for this.  You can simply edit the metadata.
- [ ] Ensure that the second author's name is represented correctly as it currently is not fully expressed as it is in the paper.

In the paper:

- [ ] LINE 126:  You are missing the DOI for this paper, which I believe may be:  [10.18637/jss.v009.i02](https://doi.org/10.18637/jss.v009.i02).  Also, please capitalize the "a" in Anderson and "d" in Darling.  You can do this in your bib file by placing curly brackets around characters you wish to maintain formatting of.
- [ ] LINE 130:  The "p" in "python" should be capitalized.

Let me know when these changes have been made and I'll accept.
