+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{marsaglia2004evaluating,
-  title   = {Evaluating the anderson-darling distribution},
+  title   = {Evaluating the Anderson-Darling distribution},
   author  = {Marsaglia, George and Marsaglia, John},
   journal = {Journal of statistical software},
   volume  = {9},
@@ -24,7 +24,7 @@
 }
 
 @article{sinclair1988approximations,
-  title     = {Approximations to the distribution function of the anderson—darling test statistic},
+  title     = {Approximations to the distribution function of the Anderson—Darling test statistic},
   author    = {Sinclair, CD and Spurr, BD},
   journal   = {Journal of the American Statistical Association},
   volume    = {83},
@@ -93,7 +93,7 @@
   author  = {Taskesen, Erdogan},
   license = {MIT},
   month   = jan,
-  title   = {{distfit is a python library for probability density fitting.}},
+  title   = {{distfit is a Python library for probability density fitting.}},
   url     = {https://erdogant.github.io/distfit},
   version = {1.4.0},
   year    = {2020}
