+++ paper/paper.bib
@@ -1,10 +1,11 @@
 @article{marsaglia2004evaluating,
   title   = {Evaluating the Anderson-Darling distribution},
   author  = {Marsaglia, George and Marsaglia, John},
-  journal = {Journal of statistical software},
+  journal = {Journal of Statistical Software},
   volume  = {9},
   pages   = {1--5},
-  year    = {2004}
+  year    = {2004},
+  doi     = {10.18637/jss.v009.i02}
 }
 
 @book{walck1996hand,
