### 2025-05-31 #7625 @crvernon
@sebastianherreramonterrosa - still missing the capitalization issue fix on LINE 126
