### 2025-05-31 #7625 @sebastianherreramonterrosa
In the paper.bib is correct. I don't why it doesn't appear in capitarl letters.
