### 2025-05-31 fcf27e6
capitalize Anderson Darling distribution
