+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{marsaglia2004evaluating,
-  title   = {Evaluating the Anderson-Darling distribution},
+  title   = {Evaluating the {A}nderson-{D}arling distribution},
   author  = {Marsaglia, George and Marsaglia, John},
   journal = {Journal of Statistical Software},
   volume  = {9},
@@ -43,7 +43,7 @@
 }
 
 @article{lewis1961distribution,
-  title     = {Distribution of the Anderson-Darling statistic},
+  title     = {Distribution of the {A}nderson-{D}arling statistic},
   author    = {Lewis, Peter AW},
   journal   = {The Annals of Mathematical Statistics},
   pages     = {1118--1124},
