### 2025-05-31 0971210
capitalize Anderson Darling references
