+++ paper/paper.bib
@@ -25,7 +25,7 @@
 }
 
 @article{sinclair1988approximations,
-  title     = {Approximations to the distribution function of the Anderson—Darling test statistic},
+  title     = {Approximations to the distribution function of the {A}nderson-{D}arling test statistic},
   author    = {Sinclair, CD and Spurr, BD},
   journal   = {Journal of the American Statistical Association},
   volume    = {83},
